% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{add_attr_to_dend}
\alias{add_attr_to_dend}
\title{Add an attribute to dendrogram nodes}
\usage{
add_attr_to_dend(dend, value, attribute = "label")
}
\arguments{
\item{dend}{dendrogram of annotated cell types (with labeled nodes)}

\item{value}{named vector of attributes to add (values = attribute value; name = dendrogram node labels)}

\item{attribute}{name of attribute to add or update in "dend"}
}
\value{
The input dendrogram with the new added/updated attribute.
}
\description{
Adds a specified attribute to labeled nodes of a dendrogram
}
