% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ccn.R
\name{as_ccn}
\alias{as_ccn}
\alias{as_ccn.default}
\alias{as_ccn.character}
\alias{as_ccn.ccn}
\alias{as_ccn.ccnr}
\title{Convert to a \code{ccn} object}
\usage{
as_ccn(x, ...)

\method{as_ccn}{default}(x, ...)

\method{as_ccn}{character}(x, ...)

\method{as_ccn}{ccn}(x, ...)

\method{as_ccn}{ccnr}(x, ...)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}

\item{...}{Passed on to methods.}
}
\value{
An S3 vector of class \verb{<ccn>}
}
\description{
Convert to a \code{ccn} object
}
\examples{
tibble::tibble(
  x = get_pin("ccn"),
  ccn = as_ccn(x),
  ccnr = as_ccnr(ccn),
  ccn2 = as_ccn(ccnr)
)
}
