% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Convert to a data.frame}
\usage{
as_data_frame(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
data.frame
}
\description{
Convert various codes to their associated names.
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
x = list("670055", "21034E", "01L008", "01J008", "05P001", "21U101", "21TA26", "45D0634589")

purrr::map(x, \(x) as.data.frame(parse(x))) |> purrr::list_rbind()

purrr::map(x, \(x) as_data_frame(x)) |> purrr::list_rbind()

collapse::join(
  purrr::map(x, \(x) as.data.frame(parse(x))) |>
  purrr::list_rbind() |>
  rlang::set_names(c("class", "ccn", "cd_ste", "cd_seq", "cd_typ", "cd_par")),
  purrr::map(x, \(x) as_data_frame(x)) |>
  purrr::list_rbind(),
  verbose = 0L)
\dontshow{\}) # examplesIf}
}
