% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr_state.R
\name{ccn_state}
\alias{ccn_state}
\title{Create a state vector}
\usage{
ccn_state(x = character())
}
\arguments{
\item{x}{A character vector of state codes}
}
\value{
an S3 vector of class \code{ccn_state}
}
\description{
A state vector here is one where all entries are two-character scalars that
correspond to a US state or territory code as defined by CMS.
}
\examples{
ccn_state(c("00", "01", "02", "72", "78", "99"))
}
