% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccn_state_codes}
\alias{ccn_state_codes}
\title{CCN State Codes}
\format{
\subsection{\code{ccn_state_codes}}{

A data frame with 108 rows and 3 columns:
\describe{
\item{state_code}{Country name}
\item{state_abbr}{2 & 3 letter ISO country codes}
\item{state_name}{Year}
...
}
}
}
\source{
\url{https://www.cms.gov/regulations-and-guidance/guidance/manuals/downloads/som107c02.pdf}
}
\usage{
ccn_state_codes
}
\description{
From Section 2779A1 – CCN for Medicare Providers
(Rev. 226; Issued: 12-06-24; Effective: 12-06-24; Implementation: 12-06-24)
}
\keyword{datasets}
