% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{cell_assignment_from_dendrogram}
\alias{cell_assignment_from_dendrogram}
\title{Assign cells to cell sets from dendrogram}
\usage{
cell_assignment_from_dendrogram(
  dend,
  samples,
  cell_id,
  mapping = data.frame(sample_name = samples),
  continue = TRUE
)
}
\arguments{
\item{dend}{dendrogram of annotated cell types}

\item{samples}{a character vector of unique sample (e.g., cell or nucleus) names}

\item{cell_id}{a character vector of cell_set_accession_ids that corresponds to each sample}

\item{mapping}{do not set this parameter}

\item{continue}{do not set this parameter}
}
\value{
A data frame where the first columns corresponds to the cell sample_name and the remaining columns correspond to cell sets from the dendrogram. Entries are either 0 = cell unassigned to cell set or 1 = cell assigned to cell set.
}
\description{
Automatically link each cell to each cell set that is available in the dendrogram to
  produce a table of the probabilities of each cell mapping to each cell type.  In this
  case we define hard probabilities (0 = unassigned to cell set; 1 = assigned to cell
  set) but this could be adapted to reflect real probabilities calculated elsewhere.
}
