% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{cell_assignment_from_groups_of_cell_types}
\alias{cell_assignment_from_groups_of_cell_types}
\title{Assign cells to cell sets from nomenclature table}
\usage{
cell_assignment_from_groups_of_cell_types(
  updated_nomenclature,
  cell_id,
  mapping,
  verbose = TRUE
)
}
\arguments{
\item{updated_nomenclature}{nomenclature table of annotated cell types}

\item{cell_id}{a character vector of cell_set_accession_ids that corresponds to each sample}

\item{mapping}{A data frame where the first columns corresponds to the cell sample_name and the remaining columns correspond to cell sets (e.g., the output from `cell_assignment_from_dendrogram`)}

\item{verbose}{A logical indicating whether to output cell set accession IDs of new annotations to screen (default=TRUE)}
}
\value{
An updated "mapping" variable with additional cells sets added from nomenclature table. Entries are either 0 = cell unassigned to cell set or 1 = cell assigned to cell set.
}
\description{
Automatically link each cell to each cell set that is included in a nomenclature table
  to produce a table of the probabilities of each cell mapping to each cell type.  In
  this case we define hard probabilities (0 = unassigned to cell set; 1 = assigned to
  cell set) but this could be adapted to reflect real probabilities calculated
  elsewhere. Only cell sets not already included in input mapping table are added.
  NOTE: This function has some issues if more than one "first_label" is in use.
}
