% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-as_data_frame.R
\name{data_frame}
\alias{data_frame}
\title{Convert to a data.frame}
\usage{
data_frame(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
data.frame
}
\description{
Convert various codes to their associated names.
}
\examples{
x = list("670055","21034E","01L008","01J008","05P001","21U101","21TA26","45D0634589")

purrr::map(x, \(x) as.data.frame(parse(x))) |> purrr::list_rbind()

purrr::map(x, \(x) as.data.frame(decode(x))) |> purrr::list_rbind()
}
