% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_rcrd.R
\name{decode_ccnr}
\alias{decode_ccnr}
\title{Decode a \code{ccnr} object}
\usage{
decode_ccnr(x)
}
\arguments{
\item{x}{\verb{<chr>} a vector of CCNs}
}
\value{
An S3 vector of class \verb{<ccnr>}
}
\description{
Decode a \code{ccnr} object
}
\examples{
decode_ccnr(get_pin("ccn"))
}
