% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{define_child_accessions}
\alias{define_child_accessions}
\title{Label cell set children}
\usage{
define_child_accessions(nomenclature)
}
\arguments{
\item{nomenclature}{the nomenclature table output from `build_nomenclature_table` or related/downstream functions}
}
\value{
An updated nomenclature table with a "child_cell_set_accessions" column appended
}
\description{
Create an additional tag called child_cell_set_accessions, which is a “|”-separated character
  vector indicating all of the child set sets. This is calculated by parsing the cell_set_label
  tags and is helpful for integration into downstream ontology use cases.
}
