% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{define_parent_accessions}
\alias{define_parent_accessions}
\title{Label parent accessions}
\usage{
define_parent_accessions(nomenclature)
}
\arguments{
\item{nomenclature}{the nomenclature table output from `build_nomenclature_table` or related/downstream functions.}
}
\value{
An updated nomenclature table with a "parent_set_accession" column appended (and also a "child_cell_set_accessions" if it wasn't already calculated)
}
\description{
Create an additional tag called parent_cell_set_accession, which is defined as the node with the fewest
children that contains a cell set.
}
\details{
Notes:

If the cell sets follows a strict hierarchy (e.g., a tree) then the parent will always be the node in the
hierarchy directly above it; however, this function still can return parents in more complex, multi-inheritance
structured tables.

This can be helpful for integration into downstream ontology use cases or can be ignored.

child_cell_set_accessions is a prerequisite for this function, and these will be calculate herein if they
haven't yet been calculated.
}
