% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{dend_to_list}
\alias{dend_to_list}
\title{Convert R dendrogram to a list}
\usage{
dend_to_list(dend, omit_names = c("markers", "markers.byCl", "class"))
}
\arguments{
\item{dend}{any R dendrogram of annotated cell types}

\item{omit_names}{character vector of attributes to exclude from the conversion to list format.  This is necessary because attributes in complex formats sometimes cannot be properly converted without the function crashing}
}
\value{
The R dendgram information in list format, with all but the omitted attributes
}
\description{
Converts an R dendrogram to a list format.  This step is necessary for outputting
  a dendrogram in json format. NOTE: If this function crashes, the "omit_names"
  variable may need to be updated to exclude additional variables in dend.
}
