% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excluded.R
\name{eipps}
\alias{eipps}
\alias{eipps_hospital}
\title{IPPS-Excluded Hospital Units/Swing-Bed Approvals}
\usage{
eipps_hospital(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare_provider("210101")
eipps_hospital("21T101")
eipps_hospital("21S101")
eipps_hospital("21U101")
}
