% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eipps_ranges}
\alias{eipps_ranges}
\title{IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals}
\format{
\subsection{\code{eipps_ranges}}{

A data frame with 3 rows and 3 columns:
\describe{
\item{start}{IPPS-Excluded range start}
\item{end}{IPPS-Excluded range end}
\item{range}{IPPS-Excluded range label}
\item{abbr}{IPPS-Excluded abbreviation}
\item{desc}{IPPS-Excluded description}
...
}
}
}
\usage{
eipps_ranges
}
\description{
IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals
}
\keyword{internal}
