% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eipps_types}
\alias{eipps_types}
\title{IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals}
\format{
\subsection{\code{eipps_types}}{

A data frame with 3 rows and 3 columns:
\describe{
\item{code}{IPPS-Excluded type code}
\item{abbr}{IPPS-Excluded type abbreviation}
\item{desc}{IPPS-Excluded type description}
...
}
}
}
\usage{
eipps_types
}
\description{
IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals
}
\keyword{internal}
