% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eipps_unit_swing_types}
\alias{eipps_unit_swing_types}
\title{IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals}
\format{
\subsection{\code{eipps_unit_swing_types}}{

A data frame with 8 rows and 3 columns:
\describe{
\item{code}{IPPS-Excluded type code}
\item{abbr}{IPPS-Excluded type abbreviation}
\item{desc}{IPPS-Excluded type description}
...
}
}
}
\source{
\url{https://www.cms.gov/regulations-and-guidance/guidance/manuals/downloads/som107c02.pdf}
}
\usage{
eipps_unit_swing_types
}
\description{
\itemize{
\item Section \strong{2779C}
\item Special Numbering System for Units of Hospitals That Are Excluded From
\item the Inpatient Prospective Payment System (IPPS) and Hospitals and CAHs
\item with Swing-Bed Approval
\item Pgs. 408-409
\item Rev. 198
\item Issued/Effective/Implementation: 01-17-20
}
}
\keyword{internal}
