% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergency.R
\name{emergency}
\alias{emergency}
\alias{is_type_emergency}
\alias{emergency_abbr}
\alias{emergency_desc}
\alias{emergency_type}
\alias{emergency_sequence}
\title{Emergency Hospital}
\usage{
is_type_emergency(x)

emergency_abbr(x)

emergency_desc(x)

emergency_type(x)

emergency_sequence(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
is_type_emergency(c("E", "F"))
emergency_abbr(c("E", "F"))
emergency_desc(c("E", "F"))
emergency_type("E")
}
