% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{emergency_code}
\alias{emergency_code}
\title{Emergency Code Lookup
Convert emergency codes to standard emergency names.}
\usage{
emergency_code(x)
}
\arguments{
\item{x}{character vector of emergency codes.}
}
\value{
character vector of emergency names.
}
\description{
Emergency Code Lookup
Convert emergency codes to standard emergency names.
}
\examples{
emergency_code(c("E", "F"))
}
