% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier_emergency.R
\name{emergency_hospital_code}
\alias{emergency_hospital_code}
\title{Emergency Hospital Code Lookup
Convert emergency codes to standard emergency names.}
\usage{
emergency_hospital_code(x)
}
\arguments{
\item{x}{character vector of emergency codes.}
}
\value{
character vector of emergency names.
}
\description{
Emergency Hospital Code Lookup
Convert emergency codes to standard emergency names.
}
\examples{
emergency_hospital_code(c("E", "F"))
}
