% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier_emergency.R
\name{emergency_hospital_name}
\alias{emergency_hospital_name}
\title{Emergency Hospital Codes}
\usage{
emergency_hospital_name(x)
}
\arguments{
\item{x}{character vector of emergency codes.}
}
\value{
character vector of emergency code names.
}
\description{
Convert emergency codes to standard emergency names.
}
\examples{
emergency_hospital_name(c("E", "F"))

}
