% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R, R/excluded.R
\name{is_ipps_excluded_type}
\alias{is_ipps_excluded_type}
\alias{excluded}
\alias{excluded_abbr}
\alias{excluded_desc}
\alias{excluded_type}
\title{IPPS Excluded Types}
\usage{
is_ipps_excluded_type(x)

excluded_abbr(x)

excluded_desc(x)

excluded_type(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
x <- c("M", "R", "S", "T", "U", "W", "Y", "Z")
excluded_abbr(x)
excluded_desc(c("E", "F"))
excluded_type("E")
}
