% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{expand_cell_set_labels}
\alias{expand_cell_set_labels}
\title{Expand cell set labels}
\usage{
expand_cell_set_labels(cell_set_label_vector, sep = " ", remerge = TRUE)
}
\arguments{
\item{cell_set_label_vector}{character vector of cell set labels for the form "{prefix} {value}"}

\item{sep}{delimiter separating prefix from value (default of " " should likely not be changed)}

\item{remerge}{should an optimally re-merged vector of cell set labels (TRUE; default) or a list of all separate cell set labels (FALSE) be returned}
}
\value{
character list or vector corresponding to desired output
}
\description{
Takes as input a vector of merged cell set labels and outputs a list of all of the cell set
  labels as vectors.  If remerge=TRUE, then instead a vector of remerged cell set labels is
  output.  Note: this function assumes a single first_label value for all cell sets.
}
