% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{find_ontology_terms}
\alias{find_ontology_terms}
\title{Find ontology terms by querying the Ontology Lookup Service directly from R. This is useful
  for quickly finding UBERON (or other) ontology IDs for inclusion in the CCN.
  Note: this is a wrapper for functions from library `rols` (http://lgatto.github.com/rols/)
  Please cite as specified in `citation("rols")` if you use this function.}
\usage{
find_ontology_terms(query, exact = TRUE, ontology = "UBERON", ...)
}
\arguments{
\item{query}{Term you want to search for in the ontology (e.g., "neocortex")}

\item{exact}{Return only exact matches to input term (TRUE; default) or that's that are close}

\item{ontology}{Which ontology to search. Default is "UBERON".  "" will search all ontologies}

\item{...}{Additional parameters to `OlsSearch`}
}
\value{
what is returned
\describe{  # Describe is optional and can go after and param or return
  \item{name}{description}
}
}
\description{
Find ontology terms by querying the Ontology Lookup Service directly from R. This is useful
  for quickly finding UBERON (or other) ontology IDs for inclusion in the CCN.
  Note: this is a wrapper for functions from library `rols` (http://lgatto.github.com/rols/)
  Please cite as specified in `citation("rols")` if you use this function.
}
