% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_codes.R
\name{get_state_name}
\alias{get_state_name}
\title{State Code Lookup
Convert state codes to standard state names.}
\usage{
get_state_name(x)
}
\arguments{
\item{x}{character vector of state codes.}
}
\value{
character vector of state names.
}
\description{
State Code Lookup
Convert state codes to standard state names.
}
\examples{
get_state_name(c("00", "01", "83", "A0"))
}
