% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_ccnr}
\alias{index_ccnr}
\title{Index a \code{ccnr} object}
\usage{
index_ccnr(x)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}
}
\value{
An S3 vector of class \verb{<ccnr_index>}
}
\description{
Index a \code{ccnr} object
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
index_ccnr(get_pin("ccn"))
\dontshow{\}) # examplesIf}
}
\keyword{internal}
