% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Init CCN Parser}
\usage{
init(num = character(0), ext = NA_character_)
}
\arguments{
\item{num}{ccn code as character vector.}

\item{ext}{optional extension code as character vector.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
init("670055")
init("05P001")
init("12345E")
init("10C0001062")
init("A5J508")
init("21T101")
init("02TA01")
}
