% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{is_ccn}
\alias{is_ccn}
\title{Is \code{x} a \code{ccn}?}
\usage{
is_ccn(x)
}
\arguments{
\item{x}{\verb{<chr>} vector}
}
\value{
\verb{<lgl>} vector
}
\description{
Is \code{x} a \code{ccn}?
}
\keyword{internal}
