% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{is_ccnr}
\alias{is_ccnr}
\title{Is \code{x} a \code{ccnr}?}
\usage{
is_ccnr(x)
}
\arguments{
\item{x}{\verb{<chr>} vector}
}
\value{
\verb{<lgl>} vector
}
\description{
Is \code{x} a \code{ccnr}?
}
\keyword{internal}
