% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_codes.R
\name{is_ipps_excluded_type}
\alias{is_ipps_excluded_type}
\title{IPPS-Excluded Types}
\usage{
is_ipps_excluded_type(x)
}
\arguments{
\item{x}{character vector}
}
\value{
logical vector
}
\description{
Identify an IPPS-excluded type
}
\examples{
is_ipps_excluded_type(LETTERS)
}
