% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_codes.R
\name{is_medicaid_facility_type}
\alias{is_medicaid_facility_type}
\title{Medicaid Facility Types}
\usage{
is_medicaid_facility_type(x)
}
\arguments{
\item{x}{A character vector, length 1}
}
\value{
A logical vector
}
\description{
Identify a Medicaid facility type
}
\examples{
is_medicaid_facility_type(LETTERS)
}
