% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier_emergency.R
\name{is_supplier_code}
\alias{is_supplier_code}
\title{Supplier Types}
\usage{
is_supplier_code(x)
}
\arguments{
\item{x}{A character vector, length 1}
}
\value{
A logical vector
}
\description{
Identify a Medicaid supplier type
}
\examples{
is_supplier_code(LETTERS)
}
