% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{make_preferred_alias_unique}
\alias{make_preferred_alias_unique}
\title{Make a unique preferred alias}
\usage{
make_preferred_alias_unique(nomenclature)
}
\arguments{
\item{nomenclature}{the nomenclature table output from `build_nomenclature_table` or related/downstream functions.}
}
\value{
A unique preferred alias
}
\description{
This is an internal function that converts the preferred_alias vector to a unique vector
 combining preferred alias and cell set label
}
