% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-medicaid.R
\name{medicaid}
\alias{medicaid}
\title{Medicaid-Only Providers}
\usage{
medicaid(x)
}
\arguments{
\item{x}{\verb{<character>} 6-position alphanumeric code.}
}
\value{
S7 object of class \verb{<Medicaid>}.
}
\description{
CCNs assigned to Medicaid-only hospitals and facilities.
}
\details{
Title XIX-only providers are identified by a 6-position alphanumeric CCN.

Positions one and two identify the state in which the provider is located.

Position three, an alpha character, identifies the type of facility, by
either level or type of care being provided.

Positions four, five and six represent a sequence beginning with \code{001}.
}
\examples{
medicaid("01L008")
medicaid("22B201")
medicaid("23E301")
medicaid("34F401")
medicaid("01J008")
medicaid("A5J508")
}
