% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaid_facility.R, R/medicaid_hospital.R
\name{medicaid_facility}
\alias{medicaid_facility}
\alias{is_type_medicaid_only_facility}
\alias{medicaid_only_facility_type_abbr}
\alias{medicaid_only_facility_type_desc}
\alias{medicaid_only_facility_sequence}
\alias{medicaid_only_facility_type}
\alias{new_medicaid_only_facility}
\alias{is_type_medicaid_hospital}
\title{Medicaid-Only Facilities}
\usage{
is_type_medicaid_only_facility(x)

medicaid_only_facility_type_abbr(x)

medicaid_only_facility_type_desc(x)

medicaid_only_facility_sequence(x)

medicaid_only_facility_type(x)

new_medicaid_only_facility(x)

is_type_medicaid_hospital(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
new_medicaid_only_facility("01L008")
new_medicaid_only_facility("22B201")
new_medicaid_only_facility("23E301")
new_medicaid_only_facility("34F401")
}
