% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaid_hospital.R
\name{medicaid_hospital}
\alias{medicaid_hospital}
\alias{medicaid_hospital_range}
\alias{medicaid_hospital_range_abbr}
\alias{medicaid_hospital_range_desc}
\alias{medicaid_only_hospital_sequence}
\alias{medicaid_only_hospital_type}
\alias{new_medicaid_only_hospital}
\title{Medicaid-Only Hospitals}
\usage{
medicaid_hospital_range(x)

medicaid_hospital_range_abbr(x)

medicaid_hospital_range_desc(x)

medicaid_only_hospital_sequence(x)

medicaid_only_hospital_type(x)

new_medicaid_only_hospital(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
new_medicaid_only_hospital("01J008")
new_medicaid_only_hospital("A5J508")
}
