% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaid_facility.R
\name{medicaid_only_facility}
\alias{medicaid_only_facility}
\title{Medicaid-Only Facilities}
\usage{
medicaid_only_facility(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{MedicaidOnlyFacility}.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicaid_only_facility("01L008")
medicaid_only_facility("22B201")
medicaid_only_facility("23E301")
medicaid_only_facility("34F401")
}
