% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicaid_only_facility_types}
\alias{medicaid_only_facility_types}
\title{CCN Medicaid-Only Facility Types}
\format{
\subsection{\code{medicaid_only_facility_types}}{

A data frame with 8 rows and 3 columns:
\describe{
\item{code}{Medicaid-only facility type code}
\item{abbr}{Medicaid-only facility type abbreviation}
\item{desc}{Medicaid-only facility type description}
...
}
}
}
\source{
\url{https://www.cms.gov/regulations-and-guidance/guidance/manuals/downloads/som107c02.pdf}
}
\usage{
medicaid_only_facility_types
}
\description{
\itemize{
\item Section \strong{2779B}
\item Title: "CMS Certification Numbers for Medicaid Providers"
\item Pgs. 407-408
\item Rev. 123
\item Issued/Effective/Implementation: 10-03-14
}
}
\keyword{internal}
