% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicaid_only_hospital_ranges}
\alias{medicaid_only_hospital_ranges}
\title{CCN Medicaid-Only Hospital Ranges}
\format{
\subsection{\code{medicaid_only_hospital_ranges}}{

A data frame with 7 rows and 3 columns:
\describe{
\item{range}{Medicaid-only hospital range}
\item{abbr}{Medicaid-only hospital abbreviation}
\item{desc}{Medicaid-only hospital description}
...
}
}
}
\usage{
medicaid_only_hospital_ranges
}
\description{
\itemize{
\item Section: \strong{2779B}
\item Title: "CMS Certification Numbers for Medicaid Providers"
\item Pgs. 407-408
\item Rev. 123
\item Iss./Eff./Impl. 10-03-14
}
}
\keyword{datasets}
