% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicaid_types}
\alias{medicaid_types}
\title{CCN Medicaid-Only Facility Types}
\format{
\subsection{\code{medicaid_types}}{

A data frame with 10 rows and 3 columns:
\describe{
\item{code}{Medicaid-only facility type code}
\item{abbr}{Medicaid-only facility type abbreviation}
\item{desc}{Medicaid-only facility type description}
...
}
}
}
\usage{
medicaid_types
}
\description{
CCN Medicaid-Only Facility Types
}
\keyword{internal}
