% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare.R
\name{medicare}
\alias{medicare}
\alias{medicare_range}
\alias{medicare_range_abbr}
\alias{medicare_range_desc}
\alias{medicare_sequence}
\title{Medicare Provider Ranges}
\usage{
medicare_range(x)

medicare_range_abbr(x)

medicare_range_desc(x)

medicare_sequence(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare_range("9000")
medicare_range_abbr("9000-9799")
medicare_range_desc("9000-9799")
medicare_sequence("9000")
}
