% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opo.R
\name{medicare_opo}
\alias{medicare_opo}
\title{Organ Procurement Organizations (OPO)}
\usage{
medicare_opo(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{MedicareOPO}.
}
\description{
Organ Procurement Organizations (OPO) based on their codes.
}
\examples{
medicare_opo("05P001")
}
