% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare.R
\name{medicare_provider}
\alias{medicare_provider}
\title{Medicare Provider}
\usage{
medicare_provider(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{Medicare}.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare_provider("670055")
}
