% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicare_ranges}
\alias{medicare_ranges}
\title{CCN Medicare Facility Ranges}
\format{
\subsection{\code{medicare_ranges}}{

A data frame with 43 rows and 6 columns:
\describe{
\item{start}{Medicare facility range start}
\item{end}{Medicare facility range end}
\item{range}{Medicare facility range label}
\item{abbr}{Medicare facility abbreviation}
\item{desc}{Medicare facility description}
\item{iv}{Medicare facility range \href{https://github.com/DavisVaughan/ivs}{{ivs::iv}} vector}
...
}
}
}
\usage{
medicare_ranges
}
\description{
CCN Medicare Facility Ranges
}
\keyword{internal}
