% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{medicare_supplier}
\alias{medicare_supplier}
\title{Medicare Supplier}
\usage{
medicare_supplier(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{Supplier}.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare_supplier("10C0001062")
medicare_supplier("45D0634589")
medicare_supplier("21X0009807")
medicare_supplier("12C0001062")
}
