% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{merge_cell_set_labels}
\alias{merge_cell_set_labels}
\title{Merge cell set labels}
\usage{
merge_cell_set_labels(cell_set_label_vector, sep = " ")
}
\arguments{
\item{cell_set_label_vector}{character vector of cell set labels for the form "{prefix} {value}"}

\item{sep}{delimiter separating prefix from value (default of " " should likely not be changed)}
}
\value{
character corresponding to merged value
}
\description{
Takes as input a vector of cell set labels and outputs a single character that merges
  the cell set labels together in a specific format.
}
