% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{overwrite_dend_node_labels}
\alias{overwrite_dend_node_labels}
\title{Overwrite dendrogram node labels}
\usage{
overwrite_dend_node_labels(dend, n = 1, lab = labels(dend))
}
\arguments{
\item{dend}{any R dendrogram of annotated cell types}

\item{n}{starting value for numbering (default of 1 does not need to be changed)}

\item{lab}{leaf labels to write, defaults to existing leaf labels}
}
\value{
R dendrogram with new node (and potentially leaf) labels
}
\description{
Overwrites node (but not leaf) labels of a dendrogram with unique labels of the
  format "n##".  This is useful for visualization of the dendrogram for manual
  annotation of cell sets.  Optionally also change leaf labels.
}
