% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{parent}
\alias{parent}
\alias{parent_unit}
\alias{parent_subunit}
\alias{has_parent}
\alias{has_parent.ccn}
\title{Assemble Unit/Subunit Parent CCN}
\usage{
parent_unit(x)

parent_subunit(x)

has_parent(x)

\method{has_parent}{ccn}(x)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}
}
\value{
A \verb{<chr>} vector of parent CCNs
}
\description{
Assemble Unit/Subunit Parent CCN
}
\examples{
parent_unit(c("45S788", "25U050", "39S027", "34S155", "29Z307"))
parent_subunit(c("21SA26", "21TA26", "02TA01", "04SD38", "52TA05"))

x = ccn(get_pin("ccn"))

tibble::tibble(
  unit = has_parent(x)$Unit,
  parent = parent_unit(has_parent(x)$Unit)) |>
  collapse::na_omit()

tibble::tibble(
  subunit = has_parent(x)$Subunit,
  parent = parent_subunit(has_parent(x)$Subunit))
}
