% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{plot_dend}
\alias{plot_dend}
\title{Plot dendrogram}
\usage{
plot_dend(dend, dendro_data = NULL, node_size = 1, r = c(-0.1, 1))
}
\arguments{
\item{dend}{any R dendrogram of annotated cell types}

\item{dendro_data}{(I don't know)}

\item{node_size}{size of nodes (default = 1)}

\item{r}{(I don't know)}
}
\value{
A plot of the dendrogram in `ggplot2` format
}
\description{
Function for plotting a dendrogram in a way that displays node labels.  This function
  is copied from `scrattch.hicat`: https://github.com/AllenInstitute/scrattch.hicat
}
