% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates}
\alias{predicates}
\alias{is_medicaid_facility_type}
\alias{is_ipps_excluded_type}
\alias{is_emergency_hospital_type}
\alias{is_supplier_type}
\title{Logical Predicates}
\usage{
is_medicaid_facility_type(x)

is_ipps_excluded_type(x)

is_emergency_hospital_type(x)

is_supplier_type(x)
}
\arguments{
\item{x}{character vector of codes to test.}
}
\value{
logical vector indicating if codes are of the specified type.
}
\description{
Logical tests for various code types
}
\examples{
x <- LETTERS
x[is_medicaid_facility_type(x)]
x[is_ipps_excluded_type(x)]
x[is_emergency_hospital_type(x)]
x[is_supplier_type(x)]
}
