% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode-location}
\alias{recode-location}
\alias{recode_state}
\alias{recode_region}
\title{Recode CMS State/Region Codes}
\usage{
recode_state(x, as = c("abbr", "full"))

recode_region(x, as = c("code", "roman", "abbr", "full"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN state codes.}

\item{as}{\verb{<chr>} format to return; one of:

For \code{recode_state()}:
\itemize{
\item \code{"abbr"}: state abbreviation (default)
\item \code{"full"}: state name
}

For \code{recode_region()}:
\itemize{
\item \code{"code"}: CMS region code (default)
\item \code{"roman"}: CMS region code as a roman numeral
\item \code{"abbr"}: CMS region's office location abbreviation
\item \code{"full"}: CMS region's office location name
}}
}
\value{
\verb{<chr>} vector of CMS state/region/office names/abbreviations
}
\description{
Recode CMS State/Region Codes
}
\examples{
tibble::tibble(
  x = ccn::states[["code"]],
  state = recode_state(x, "abbr"),
  region = recode_region(state, "code"),
  office = recode_region(state, "full")
)
}
