% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode-types}
\alias{recode-types}
\alias{recode_medicaid_type}
\alias{recode_unit_type}
\alias{recode_subunit_type}
\alias{recode_other_type}
\title{Recode CMS Facility Types}
\usage{
recode_medicaid_type(x, as = c("abbr", "full"))

recode_unit_type(x, as = c("abbr", "full", "infix", "eipps"))

recode_subunit_type(x, as = c("prefix", "abbr"))

recode_other_type(x, as = c("abbr", "full"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN facility type characters.}

\item{as}{\verb{<chr>} format to return; one of:

For \code{recode_medicaid_type()}:
\itemize{
\item \code{"abbr"}: facility type abbreviation (default)
\item \code{"full"}: facility type name
}

For \code{recode_unit_type()}:
\itemize{
\item \code{"abbr"}: unit abbreviation (default)
\item \code{"full"}: unit name
\item \code{"infix"}: unit parent ccn infix
\item \code{"eipps"}: is type IPPS-excluded?
}

For \code{recode_subunit_type()}:
\itemize{
\item \code{"prefix"}: subunit parent ccn infix (default)
\item \code{"abbr"}: subunit abbreviation
}

For \code{recode_other_type()}:
\itemize{
\item \code{"abbr"}: facility type abbreviation (default)
\item \code{"full"}: facility type name
}}
}
\value{
\verb{<chr>} vector of facility type names/abbreviations
}
\description{
Recode CMS Facility Types
}
\examples{
tibble::tibble(
  x = ccn::medicaid_types[["code"]],
  abbr = recode_medicaid_type(x, "abbr"),
  full = recode_medicaid_type(x, "full")
)

tibble::tibble(
  x = ccn::unit_types[["code"]],
  abbr = recode_unit_type(x, "abbr"),
  full = recode_unit_type(x, "full"),
  infix = recode_unit_type(x, "infix"),
  eipps = recode_unit_type(x, "eipps"),
)

tibble::tibble(
  x = ccn::subunit_types[["code"]],
  prefix = recode_subunit_type(x, "prefix"),
  abbr = recode_subunit_type(x, "abbr")
)

tibble::tibble(
  x = c("C", "D", "E", "F", "P", "X"),
  abbr = recode_other_type(x, "abbr"),
  full = recode_other_type(x, "full")
)
}
