% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_medicaid_type}
\alias{recode_medicaid_type}
\title{Recode Medicaid Facility Types}
\usage{
recode_medicaid_type(x, as = c("abbr", "full"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN Medicaid facility type characters.}

\item{as}{\verb{<chr>} format to return; one of:
\itemize{
\item "abbr": state abbreviation (default)
\item "full": state name
}}
}
\value{
\verb{<chr>} vector of Medicaid facility type names/abbreviations
}
\description{
Recode Medicaid Facility Types
}
\examples{
tibble::tibble(
  code = c("A", "K", NA),
  abbr = recode_medicaid_type(code, "abbr"),
  full = recode_medicaid_type(code, "full")
)
}
