% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_other_type}
\alias{recode_other_type}
\title{Recode Other Facility Types}
\usage{
recode_other_type(x, as = c("abbr", "full"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN Other facility type characters.}

\item{as}{\verb{<chr>} format to return; one of:
\itemize{
\item \code{"abbr"}: facility type abbreviation (default)
\item \code{"full"}: facility type name
}}
}
\value{
\verb{<chr>} vector of Other facility type names/abbreviations
}
\description{
Recode Other Facility Types
}
\examples{
tibble::tibble(
  x = c("C", "D", "E", "F", "P", "X", NA_character_),
  abbr = recode_other_type(x, "abbr"),
  full = recode_other_type(x, "full")
)
}
