% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_region}
\alias{recode_region}
\title{Recode CMS Region Codes}
\usage{
recode_region(x, as = c("number", "roman", "abbr", "full"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN state codes.}

\item{as}{\verb{<chr>} format to return; one of:
\itemize{
\item \code{"number"}: CMS region number (default)
\item \code{"roman"}: CMS region number as a roman numeral
\item \code{"abbr"}: CMS region's office location abbreviation
\item \code{"full"}: CMS region's office location name
}}
}
\value{
\verb{<chr>} vector of CMS region/region office names/abbreviations
}
\description{
Recode CMS Region Codes
}
\examples{
tibble::tibble(
  x = c("00", "01", "A5"),
  number = recode_region(x, "number"),
  roman = recode_region(x, "roman"),
  abbr = recode_region(x, "abbr"),
  full = recode_region(x, "full")
)
}
