% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_state}
\alias{recode_state}
\title{Recode State Codes}
\usage{
recode_state(x, as = c("abbr", "full"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN state codes.}

\item{as}{\verb{<chr>} format to return; one of:
\itemize{
\item \code{"abbr"}: state abbreviation (default)
\item \code{"full"}: state name
}}
}
\value{
\verb{<chr>} vector of state names/abbreviations
}
\description{
Recode State Codes
}
\examples{
tibble::tibble(
  x = c("00", "01", "A5"),
  abbr = recode_state(x, "abbr"),
  full = recode_state(x, "full")
)
}
