% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_subunit_type}
\alias{recode_subunit_type}
\title{Recode Subunit Facility Types}
\usage{
recode_subunit_type(x, as = c("prefix", "abbr"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN Subunit facility type characters.}

\item{as}{\verb{<chr>} format to return; one of:
\itemize{
\item \code{"prefix"}: subunit parent ccn infix (default)
\item \code{"abbr"}: subunit abbreviation
}}
}
\value{
\verb{<chr>} vector of Subunit facility type abbreviations/parent prefixes
}
\description{
Recode Subunit Facility Types
}
\examples{
tibble::tibble(
  x = ccn::subunit_types[["code"]],
  prefix = recode_subunit_type(x, "prefix"),
  abbr = recode_subunit_type(x, "abbr")
)
}
