% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_unit_type}
\alias{recode_unit_type}
\title{Recode Unit Facility Types}
\usage{
recode_unit_type(x, as = c("abbr", "full", "infix", "eipps"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN Unit facility type characters.}

\item{as}{\verb{<chr>} format to return; one of:
\itemize{
\item \code{"abbr"}: unit abbreviation (default)
\item \code{"full"}: unit name
\item \code{"infix"}: unit parent ccn infix
\item \code{"eipps"}: is type IPPS-excluded?
}}
}
\value{
\verb{<chr>} vector of Unit facility type names/abbreviations/parent infixes/IPPS exclusion statuses
}
\description{
Recode Unit Facility Types
}
\examples{
tibble::tibble(
  x = c("M", "T", "Z"),
  abbr = recode_unit_type(x, "abbr"),
  full = recode_unit_type(x, "full"),
  infix = recode_unit_type(x, "infix"),
  eipps = recode_unit_type(x, "eipps"),
)
}
