% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional_offices}
\alias{regional_offices}
\title{CMS Regional Office Codes}
\format{
\subsection{\code{regional_offices}}{

A data frame with 59 rows and 4 columns:
\describe{
\item{ro}{Region number}
\item{ro_region}{Region roman numeral}
\item{state}{State abbreviation}
\item{ro_city}{Regional office city}
...
}
}
}
\usage{
regional_offices
}
\description{
CMS Regional Office Codes
}
\keyword{internal}
