% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{CMS Regional Office Codes}
\format{
\subsection{\code{regions}}{

A data frame with 109 rows and 5 columns:
\describe{
\item{code}{Region number}
\item{roman}{Region roman numeral}
\item{state}{State abbreviation}
\item{office_abbr}{Regional office city abbreviation}
\item{office_full}{Regional office city full name}
...
}
}
}
\usage{
regions
}
\description{
CMS Regional Office Codes
}
\keyword{internal}
