% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{sequence}
\alias{sequence}
\alias{sequence_medicare}
\alias{sequence_medicaid_only}
\alias{sequence_supplier}
\alias{sequence_emergency}
\alias{sequence_opo}
\title{Sequence Class and Constructors}
\usage{
sequence_medicare(x)

sequence_medicaid_only(x)

sequence_supplier(x)

sequence_emergency(x)

sequence_opo(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
sequence_medicare("0055")
sequence_medicaid_only("232")
sequence_supplier("9000000")
sequence_emergency("500")
sequence_opo("50")
}
