% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{slice}
\alias{slice}
\alias{switch_provider}
\alias{switch_ccn}
\alias{slice_provider_ext}
\alias{slice_medicare}
\alias{slice_organ}
\alias{slice_emergency}
\alias{slice_supplier}
\alias{slice_supplier_ext}
\alias{slice_medicaid}
\alias{slice_unit}
\alias{slice_subunit}
\alias{print.medicare}
\alias{print.organ}
\alias{print.emergency}
\alias{print.supplier}
\alias{print.medicaid}
\alias{print.unit}
\alias{print.subunit}
\title{Internal slice functions}
\usage{
switch_provider(x)

switch_ccn(x)

slice_provider_ext(x)

slice_medicare(x)

slice_organ(x)

slice_emergency(x)

slice_supplier(x)

slice_supplier_ext(x)

slice_medicaid(x)

slice_unit(x)

slice_subunit(x)

\method{print}{medicare}(x, ...)

\method{print}{organ}(x, ...)

\method{print}{emergency}(x, ...)

\method{print}{supplier}(x, ...)

\method{print}{medicaid}(x, ...)

\method{print}{unit}(x, ...)

\method{print}{subunit}(x, ...)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Internal slice functions
}
\keyword{internal}
