% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.R, R/method-print.R, R/slice.R
\name{as.data.frame.medicare}
\alias{as.data.frame.medicare}
\alias{as.data.frame.organ}
\alias{as.data.frame.emergency}
\alias{as.data.frame.supplier}
\alias{as.data.frame.medicaid}
\alias{as.data.frame.unit}
\alias{as.data.frame.subunit}
\alias{print.medicare}
\alias{print.organ}
\alias{print.emergency}
\alias{print.supplier}
\alias{print.medicaid}
\alias{print.unit}
\alias{print.subunit}
\alias{slice_medicare}
\alias{slice_provider_ext}
\alias{slice_supplier_ext}
\alias{slice_organ}
\alias{slice_emergency}
\alias{slice_supplier}
\alias{slice_medicaid}
\alias{slice_unit}
\alias{slice_subunit}
\title{Medicare Provider: 670055 -> 67 0055}
\usage{
\method{as.data.frame}{medicare}(x, ...)

\method{as.data.frame}{organ}(x, ...)

\method{as.data.frame}{emergency}(x, ...)

\method{as.data.frame}{supplier}(x, ...)

\method{as.data.frame}{medicaid}(x, ...)

\method{as.data.frame}{unit}(x, ...)

\method{as.data.frame}{subunit}(x, ...)

\method{print}{medicare}(x, ...)

\method{print}{organ}(x, ...)

\method{print}{emergency}(x, ...)

\method{print}{supplier}(x, ...)

\method{print}{medicaid}(x, ...)

\method{print}{unit}(x, ...)

\method{print}{subunit}(x, ...)

slice_medicare(x)

slice_provider_ext(x)

slice_supplier_ext(x)

slice_organ(x)

slice_emergency(x)

slice_supplier(x)

slice_medicaid(x)

slice_unit(x)

slice_subunit(x)
}
\arguments{
\item{x}{character vector of codes to look up.}

\item{...}{additional arguments (not used).}
}
\description{
Medicare Provider: 670055 -> 67 0055

Medicare Supplier with Extension: 10C0001062EXT -> 10 C 0001062 EXT

Medicare OPO Provider: 05P001 -> 05 P 001

Emergency Hospital: 12345E -> 12 345 E

Medicare Supplier: 10C0001062 -> 10 C 0001062

Medicaid-Only Facility: 01L008 -> 01 L 008

IPPS-Excluded Provider: 21T101 -> 21 T 101

IPPS-Excluded Unit: 02TA01 -> 02 T A 01
}
\keyword{internal}
