% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{split_column}
\alias{split_column}
\title{Split a column}
\usage{
split_column(nomenclature, column_name, split_char = "\\\\|", expand = FALSE)
}
\arguments{
\item{nomenclature}{the nomenclature table output from `build_nomenclature_table` or related/downstream functions.}

\item{column_name}{the name of the column for splitting}

\item{split_char}{the character to split the column on (default = "|")}

\item{expand}{expands the matrix (default = FALSE)}
}
\value{
A matrix with character entries, where each row corresponds to an entry from the input vector and each column corresponds to unique values from the split vectors
}
\description{
This function splits a column in the nomenclature table by a defined character into a matrix of values for use with downstream CCN functions
}
