% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{state}
\alias{state}
\title{State Codes}
\usage{
state(x)
}
\arguments{
\item{x}{character vector of state codes to look up.}
}
\value{
S7 object of class \code{State}.
}
\description{
Convert state codes to their abbreviations or full names.
}
\examples{
state("A5XXX")
state(sample(ccn::state_codes$code, 5L)) # should only accept length 1
}
