% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_code.R
\name{state_code}
\alias{state_code}
\alias{state_abbr}
\alias{state_name}
\alias{state_region}
\alias{state_office}
\alias{state}
\title{State Code Conversions}
\usage{
state_abbr(code)

state_name(code)

state_region(abbr)

state_office(abbr)

state(ccn)
}
\arguments{
\item{code}{character vector of 2-character CCN state codes.}

\item{abbr}{character vector of 2-character state abbreviations.}

\item{ccn}{6-character CCN code.}
}
\value{
S7 object of class \code{State}.
}
\description{
Convert CCN state codes to their abbreviations or full names.
Convert state abbreviations to their RO region or office (city).
}
\examples{
state_region(abbr = "CA")
state_office(abbr = "NY")
state_abbr(code = c("00", "A5", "01", "99"))
state_name(code = c("00", "A5", "01", "99"))
state(ccn = "A5XXX")
state(ccn = "869999")
}
