% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_codes}
\alias{state_codes}
\title{CCN State Codes}
\format{
\subsection{\code{state_codes}}{

A data frame with 108 rows and 3 columns:
\describe{
\item{code}{State code}
\item{abbr}{State abbreviation}
\item{name}{State full name}
...
}
}
}
\usage{
state_codes
}
\description{
CCN State Codes
}
\keyword{internal}
