% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunit.R
\name{subunit}
\alias{subunit}
\title{IPPS-Excluded Hospital Subunits}
\usage{
subunit(x)
}
\arguments{
\item{x}{\verb{<character>} 6-position alphanumeric code.}
}
\value{
S7 object of class \verb{<Subunit>}.
}
\description{
Convert various codes to their associated names.
}
\examples{
subunit("21SA26")
subunit("21TA26")
subunit("02TA01")
subunit("04SD38")
subunit("52TA05")
}
