% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subunit_types}
\alias{subunit_types}
\title{IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals}
\format{
\subsection{\code{subunit_types}}{

A data frame with 10 rows and 3 columns:
\describe{
\item{code}{IPPS-Excluded type code}
\item{prefix}{IPPS-Excluded range prefix}
\item{abbr}{IPPS-Excluded type description}
...
}
}
}
\usage{
subunit_types
}
\description{
IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals
}
\keyword{internal}
