% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{supplier}
\alias{supplier}
\title{Medicare Supplier}
\usage{
supplier(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{Supplier}.
}
\description{
Convert various codes to their associated names.
}
\examples{
supplier("10C0001062")
supplier("45D0634589")
supplier("21X0009807")
supplier("12C0001062")
}
