% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier_emergency.R
\name{supplier_code}
\alias{supplier_code}
\title{Supplier Code Lookup
Convert supplier codes to standard supplier names.}
\usage{
supplier_code(x)
}
\arguments{
\item{x}{character vector of supplier codes.}
}
\value{
character vector of supplier names.
}
\description{
Supplier Code Lookup
Convert supplier codes to standard supplier names.
}
\examples{
supplier_code(c("C", "D", "X"))
}
