% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{types}
\alias{types}
\alias{is_type_supplier}
\alias{supplier_abbr}
\alias{supplier_desc}
\alias{type_supplier}
\alias{is_type_emergency}
\alias{emergency_abbr}
\alias{emergency_desc}
\alias{type_emergency}
\alias{is_type_opo}
\alias{opo_abbr}
\alias{opo_desc}
\alias{type_opo}
\alias{is_type_medicaid_only}
\alias{medicaid_only_abbr}
\alias{medicaid_only_desc}
\alias{type_medicaid_only}
\alias{is_type_ipps_excluded}
\alias{ipps_excluded_abbr}
\alias{ipps_excluded_desc}
\alias{type_ipps_excluded}
\title{Facility Type Codes}
\usage{
is_type_supplier(x)

supplier_abbr(x)

supplier_desc(x)

type_supplier(x)

is_type_emergency(x)

emergency_abbr(x)

emergency_desc(x)

type_emergency(x)

is_type_opo(x)

opo_abbr(x)

opo_desc(x)

type_opo(x)

is_type_medicaid_only(x)

medicaid_only_abbr(x)

medicaid_only_desc(x)

type_medicaid_only(x)

is_type_ipps_excluded(x)

ipps_excluded_abbr(x)

ipps_excluded_desc(x)

type_ipps_excluded(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
x <- LETTERS

x[is_type_supplier(x)]
supplier_abbr(x[is_type_supplier(x)])
supplier_desc(x[is_type_supplier(x)])
type_supplier("C")

x[is_type_emergency(x)]
emergency_abbr(c("E", "F"))
emergency_desc(c("E", "F"))
type_emergency("E")

x[is_type_opo(x)]
opo_abbr(x[is_type_opo(x)])
opo_desc(x[is_type_opo(x)])
type_opo(x[is_type_opo(x)])

x[is_type_medicaid_only(x)]
medicaid_only_abbr(x[is_type_medicaid_only(x)])
medicaid_only_desc(x[is_type_medicaid_only(x)])
type_medicaid_only("A")

x[is_type_ipps_excluded(x)]
ipps_excluded_abbr(x[is_type_ipps_excluded(x)])
ipps_excluded_desc(x[is_type_ipps_excluded(x)])
type_ipps_excluded("M")
}
