% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{unit}
\alias{unit}
\alias{parent_unit}
\alias{parent_subunit}
\title{Assemble Unit/Subunit Parent CCN}
\usage{
parent_unit(x)

parent_subunit(x)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}
}
\value{
A \verb{<chr>} vector of parent CCNs
}
\description{
Assemble Unit/Subunit Parent CCN
}
\examples{
parent_unit(c("45S788", "25U050", "39S027", "34S155", "29Z307"))
parent_subunit(c("21SA26", "21TA26", "02TA01", "04SD38", "52TA05"))

x = ccn(get_pin("ccn"))
i = get_index(x)

tibble::tibble(
  unit = x[i$Unit],
  parent = parent_unit(x[i$Unit])) |>
  collapse::na_omit()

tibble::tibble(
  subunit = x[i$Subunit],
  parent = parent_subunit(x[i$Subunit]))
}
