% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_types}
\alias{unit_types}
\title{IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals}
\format{
\subsection{\code{unit_types}}{

A data frame with 9 rows and 4 columns:
\describe{
\item{code}{IPPS-Excluded type code}
\item{abbr}{IPPS-Excluded type abbreviation}
\item{desc}{IPPS-Excluded type description}
\item{infix}{IPPS-Excluded range infix}
\item{eipps}{IPPS-Excluded type description}
...
}
}
}
\usage{
unit_types
}
\description{
IPPS-Excluded Hospital Unit Types/Swing-Bed Approvals
}
\keyword{internal}
