% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_scripts.R
\name{update_dendrogram_with_nomenclature}
\alias{update_dendrogram_with_nomenclature}
\title{Add nomenclature annotations to dendrogram}
\usage{
update_dendrogram_with_nomenclature(
  dend,
  cell_set_information,
  current_label = "original_label",
  new_label = "cell_set_preferred_alias"
)
}
\arguments{
\item{dend}{dendrogram of cell types to annotate}

\item{cell_set_information}{any table or data frame with information to annotate a dendrogram, typically the nomenclature table output from `build_nomenclature_table`}

\item{current_label}{column name in `cell_set_information` that contains dendrogram labels (default of "original_label" should be retained if running this function after `build_nomenclature_table`)}

\item{new_label}{column name holding desired dendrogram labels after running script (default "cell_set_preferred_alias" )}
}
\value{
An annotated dendrogram with updated labels
}
\description{
This code will take the information from the table above and add it to the initial
  dendrogram object. When plotted the only visible difference will be that the new
  cell set alias names (if any) will show up to replace the n## labels from the initial
  plot. However, ALL of the meta-data read in from the table will be added to the
  relevant nodes or leafs.
}
