% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctr_state.R
\name{vctr_state}
\alias{vctr_state}
\alias{validate_state}
\alias{is_state}
\alias{as_state}
\alias{vec_ptype_abbr.state}
\alias{vec_ptype_full.state}
\alias{format.state}
\title{Create a state vector}
\usage{
vctr_state(x = character())

validate_state(x)

is_state(x)

as_state(x)

\method{vec_ptype_abbr}{state}(x, ...)

\method{vec_ptype_full}{state}(x, ...)

\method{format}{state}(x, ...)
}
\arguments{
\item{x}{A character vector of state codes}

\item{...}{Additional arguments (not used)}
}
\value{
an S3 vector of class \code{vctr_state}
}
\description{
A state vector here is one where all entries are two-character scalars that
correspond to a US state or territory code as defined by CMS.
}
\examples{
vctr_state(c("00", "01", "02", "72", "78", "99"))
}
