% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{vec_cast.ccnr}
\alias{vec_cast.ccnr}
\title{Casting}
\usage{
\method{vec_cast}{ccnr}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}}.
}
\keyword{internal}
