% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{vec_restore.ccn}
\alias{vec_restore.ccn}
\title{Restore}
\usage{
\method{vec_restore}{ccn}(x, to, ..., i = NULL)
}
\arguments{
\item{x}{A vector.}

\item{to}{The original vector to restore to.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Restore
}
\keyword{internal}
