% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as}
\alias{as}
\alias{as_ccn}
\alias{as_ccn.default}
\alias{as_ccn.character}
\alias{as_ccn.ccn}
\alias{as_ccn.ccnr}
\alias{as_ccnr}
\alias{as_ccnr.character}
\alias{as_ccnr.ccn}
\alias{as_ccnr.ccnr}
\title{Convert to a \code{ccn}/\code{ccnr} object}
\usage{
as_ccn(x, ...)

\method{as_ccn}{default}(x, ...)

\method{as_ccn}{character}(x, ...)

\method{as_ccn}{ccn}(x, ...)

\method{as_ccn}{ccnr}(x, ...)

as_ccnr(x, ...)

\method{as_ccnr}{character}(x, ...)

\method{as_ccnr}{ccn}(x, ...)

\method{as_ccnr}{ccnr}(x, ...)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}

\item{...}{Passed on to methods.}
}
\value{
An S3 vector of class \verb{<ccn>} or \verb{<ccnr>}
}
\description{
Convert to a \code{ccn}/\code{ccnr} object
}
\examples{
x = get_pin("ccn")

head(as_ccn(x))

head(as_ccnr(x))

}
