% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ccn.R
\name{as_ccnr}
\alias{as_ccnr}
\alias{as_ccnr.character}
\alias{as_ccnr.ccn}
\alias{as_ccnr.ccnr}
\title{Convert to a \code{ccnr} object}
\usage{
as_ccnr(x, ...)

\method{as_ccnr}{character}(x, ...)

\method{as_ccnr}{ccn}(x, ...)

\method{as_ccnr}{ccnr}(x, ...)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}

\item{...}{Passed on to methods.}
}
\value{
An S3 vector of class \verb{<ccnr>}
}
\description{
Convert to a \code{ccnr} object
}
\examples{
as_ccnr(get_pin("ccn"))
}
