% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{as_data_frame}
\alias{as_data_frame}
\title{Convert to data.frame}
\usage{
as_data_frame(x, ...)
}
\arguments{
\item{x}{ccn object}

\item{...}{additional arguments}
}
\value{
data.frame
}
\description{
Convert to data.frame
}
\examples{
(x <- ccn("01L008"))

as_data_frame(x) |> str()
}
