% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn-package.R
\docType{package}
\name{ccn-package}
\alias{ccn}
\alias{ccn-package}
\title{ccn: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/ccn}
  \item \url{https://andrewallenbruce.github.io/ccn/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/ccn/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
