% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn-package.R
\docType{package}
\name{ccn-package}
\alias{ccn}
\alias{ccn-package}
\title{ccn: CMS Certification Numbers}
\description{
An \href{https://github.com/RConsortium/S7/}{S7 class} representing a
\strong{CMS Certification Number}, a uniform way of identifying Providers or
Suppliers who have participated in the Medicare or Medicaid programs.

A CCN is an alphanumeric string, either 6 or 10 characters in length, that
encodes a Provider/Supplier's State and Facility type, as defined by
\href{https://www.cms.gov/Regulations-and-Guidance/Guidance/Transmittals/downloads/R3SOM.pdf}{CMS Manual System Pub. 100-07}
\emph{Medicare State Operations: Provider Certification}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/ccn}
  \item \url{https://andrewallenbruce.github.io/ccn/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/ccn/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
