% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn-package.R
\docType{package}
\name{ccn-package}
\alias{ccn-package}
\title{CCN}
\description{
An \href{https://github.com/RConsortium/S7/}{S7} class representing a \strong{CMS Certification Number},
a uniform way of identifying providers or suppliers who currently or who
ever have participated in the Medicare or Medicaid programs.
}
\details{
A CCN is a 6- or 10-character alphanumeric string that encodes the provider
or supplier's state and facility type, as defined by
\strong{CMS Manual System publication number 100-07 "State Operations Provider Certification"}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/ccn}
  \item \url{https://andrewallenbruce.github.io/ccn/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/ccn/issues}
}

}
\author{
\strong{Maintainer}: Andrew Allen Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
