% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{ccn}
\alias{ccn}
\alias{is_ccn}
\title{Construct a \code{ccn} object}
\usage{
ccn(x = character())

is_ccn(x)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}
}
\value{
An S3 vector of class \verb{<ccn>}
}
\description{
Construct a \code{ccn} object
}
\examples{
x <- get_pin("ccn")
y <- ccn(x)
is_ccn(x)
is_ccn(y)
vctrs::vec_c(x[1:5], y[100:150], x[1:50])
tibble::tibble(x = x, ccn = ccn(x))
}
