% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_vctr.R, R/index.R
\name{ccn}
\alias{ccn}
\alias{is_ccn}
\alias{index_ccn_types}
\title{Construct a \code{ccn} object}
\usage{
ccn(x = character())

is_ccn(x)

index_ccn_types(x)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}
}
\value{
An S3 vector of class \verb{<ccn>}
}
\description{
\itemize{
\item \code{new_ccn()} is a low-level constructor that takes a vector.
\item \code{ccn()} constructs an npi object from a vector.
\item \code{as_ccn()} and \code{is_ccn()} forward to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[=inherits]{inherits()}}, respectively.
}
}
\examples{
x <- get_pin("ccn")

y <- ccn(x)

is_ccn(x)
is_ccn(y)

vctrs::vec_c(x[1:5], y[100:150], x[1:50])

tibble::tibble(x = x, ccn = y)

index_ccn_types(y)

}
