% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-ccn.R, R/vctrs-infer.R
\name{ccn}
\alias{ccn}
\alias{new_ccn}
\alias{is_ccn}
\alias{as_ccn}
\alias{as_ccn.default}
\alias{as_ccn.character}
\alias{as_ccn.ccnr}
\alias{index_types}
\title{Construct a \code{ccn} object}
\usage{
ccn(x = character())

new_ccn(x = character())

is_ccn(x)

as_ccn(x, ...)

\method{as_ccn}{default}(x, ...)

\method{as_ccn}{character}(x, ...)

\method{as_ccn}{ccnr}(x, ...)

index_types(x)
}
\arguments{
\item{x}{a vector}

\item{...}{Passed on to methods.}
}
\value{
An S3 vector of class \verb{<ccn>}
}
\description{
\itemize{
\item \code{new_ccn()} is a low-level constructor that takes a vector.
\item \code{ccn()} constructs an npi object from a vector.
\item \code{as_ccn()} and \code{is_ccn()} forward to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[=inherits]{inherits()}}, respectively.
}
}
\examples{
x <- get_pin("ccn")
y <- ccn(x)
vctrs::vec_c(x[1:5], y[100:150], x[1:50])
tibble::tibble(x = x, ccn = as_ccn(x))
index_types(x)
}
