% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-ccn.R
\name{ccn}
\alias{ccn}
\alias{new_ccn}
\alias{form}
\alias{is_ccn}
\alias{as_ccn}
\alias{as_ccn.default}
\alias{as_ccn.character}
\alias{ccn_form}
\title{Construct a \code{ccn} object}
\usage{
ccn(x = character())

new_ccn(x = character(), form = character())

form(x)

is_ccn(x)

as_ccn(x, ...)

\method{as_ccn}{default}(x, ...)

\method{as_ccn}{character}(x)

ccn_form(x)
}
\arguments{
\item{x}{a vector}

\item{form}{ccn form}

\item{...}{Passed on to methods.}
}
\value{
An S3 vector of class \verb{<ccn>}
}
\description{
\itemize{
\item \code{new_ccn()} is a low-level constructor that takes a vector.
\item \code{ccn()} constructs an npi object from a vector.
\item \code{as_ccn()} and \code{is_ccn()} forward to \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}} and \code{\link[=inherits]{inherits()}}, respectively.
}
}
\examples{
\dontshow{if (FALSE) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
