% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{ccn}
\alias{ccn}
\title{Create a ccn vector}
\usage{
ccn(x = character())
}
\arguments{
\item{x}{A character vector of ccn codes}
}
\value{
an S3 vector of class \code{ccn}
}
\description{
A \code{ccn} vector here is one where all entries are character scalars with a
length between 6 and 14
}
\examples{
ccn(c("67-0055", "21034e", "01L008", "01J008", "05P001", "21U101", "21TA26", "45D0634589"))
}
