% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{ccn}
\alias{ccn}
\title{Decode a CCN}
\usage{
ccn(x)
}
\arguments{
\item{x}{character vector of CCNs.}
}
\value{
list of CCN components.
}
\description{
Decode a CCN into its component parts.
}
\examples{
# Medicare Provider
ccn("670055")
ccn("05P001") # OPO

ccn("210101")
ccn("21T101")
ccn("21S101")
ccn("21U101")

# Medicaid Only Provider
ccn("01L008")

# Emergency Hospital
ccn("12345E")

# Supplier
ccn("10C0001062") # ASC
ccn("45D0634589") # CLIA
ccn("21X0009807") # Portable X-Ray

# IPPS Excluded Unit
ccn("02TA01")
ccn("04SD38")
ccn("52TA05")

ccn("212026") # Parent
ccn("21SA26")
ccn("21TA26")
}
