% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{ccn}
\alias{ccn}
\title{Decode a CCN}
\usage{
ccn(x)
}
\arguments{
\item{x}{character vector of CCNs.}
}
\value{
list of CCN components.
}
\description{
Decode a CCN into its component parts.
}
\examples{
ccn("01P001") # OPO
ccn("05P001") # OPO
ccn("670055")

ccn("210101")
ccn("21T101")
ccn("21S101")
ccn("21U101")

ccn("212026")
ccn("21SA26")
ccn("21TA26")

ccn("01L008") # Medicaid Only Provider
ccn("12345E") # Emergency Hospital
ccn("10C0001062") # Supplier ASC
ccn("45D0634589") # Supplier CLIA
ccn("21X0009807") # Supplier Portable X-Ray

# IPPS Excluded Units
ccn("02TA01")
ccn("04SD38")
ccn("52TA05")
}
