% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{ccn}
\alias{ccn}
\title{Decode a CCN}
\usage{
ccn(x)
}
\arguments{
\item{x}{character vector of CCNs.}
}
\value{
list of CCN components.
}
\description{
Decode a CCN into its component parts.
}
\examples{
ccn("67P055")
ccn("670055")
ccn("21-0101")
ccn("21-T101")
ccn("12345E")
ccn("12C4567890")
ccn("0-12C4567890")
}
