% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_rcrd.R
\name{ccnr}
\alias{ccnr}
\title{Construct a \code{ccnr} object}
\usage{
ccnr(
  entity = character(),
  ccn = character(),
  state = character(),
  number = character(),
  type = character()
)
}
\arguments{
\item{entity}{\verb{<chr>} vector of ccn entity types}

\item{ccn}{\verb{<chr>} vector of ccns}

\item{state}{\verb{<chr>} vector of ccn state codes}

\item{number}{\verb{<chr>} vector of ccn code sequence numbers}

\item{type}{\verb{<chr>} vector of ccn code facility types}
}
\value{
An S3 vector of class \verb{<ccnr>}
}
\description{
Construct a \code{ccnr} object
}
\examples{
tibble::tibble(
  x = get_pin("ccn"),
  ccn = as_ccn(x),
  ccnr = as_ccnr(ccn))
}
