% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_rcrd.R
\name{ccnr}
\alias{ccnr}
\title{Construct a \code{ccnr} object}
\usage{
ccnr(
  entity = character(),
  ccn = character(),
  state = character(),
  number = character(),
  type = character()
)
}
\arguments{
\item{entity}{\verb{<chr>} a vector of CCN Entity types}

\item{ccn}{\verb{<chr>} A vector of CCNs}

\item{state}{\verb{<chr>} Passed on to methods}

\item{number}{\verb{<chr>} a vector}

\item{type}{\verb{<chr>} a vector}
}
\value{
An S3 vector of class \verb{<ccnr>}
}
\description{
Construct a \code{ccnr} object
}
\examples{
tibble::tibble(
  x = get_pin("ccn"),
  ccn = as_ccn(x),
  ccnr = as_ccnr(ccn))
}
