% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccnr.R
\name{ccnr}
\alias{ccnr}
\alias{is_ccnr}
\title{Construct a \code{ccnr} object}
\usage{
ccnr(
  entity = character(),
  ccn = character(),
  state = character(),
  number = character(),
  type = character()
)

is_ccnr(x)
}
\arguments{
\item{entity}{\verb{<chr>} vector of ccn entity types}

\item{ccn}{\verb{<chr>} vector of ccns}

\item{state}{\verb{<chr>} vector of ccn state codes}

\item{number}{\verb{<chr>} vector of ccn code sequence numbers}

\item{type}{\verb{<chr>} vector of ccn code facility types}

\item{x}{\verb{<chr>} A vector of CCNs.}
}
\value{
An S3 vector of class \verb{<ccnr>}
}
\description{
Construct a \code{ccnr} object
}
\examples{
ccnr()

ccnr(
  entity = "Medicare",
  ccn = "001234",
  state = "00",
  number = "1234",
  type = NA_character_
 )
}
