% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-ccnr.R
\name{ccnr}
\alias{ccnr}
\alias{new_ccnr}
\alias{is_ccnr}
\alias{as_ccnr}
\alias{decode_ccnr}
\title{Construct a \code{ccnr} object}
\usage{
ccnr(
  ccn = character(),
  entity = character(),
  state = character(),
  number = character(),
  type = character(),
  parent = character(),
  ext = character()
)

new_ccnr(
  ccn = character(),
  entity = character(),
  state = character(),
  number = character(),
  type = character(),
  parent = character(),
  ext = character()
)

is_ccnr(x)

as_ccnr(x)

decode_ccnr(x)
}
\arguments{
\item{ccn}{a vector}

\item{entity}{a vector}

\item{state}{Passed on to methods}

\item{number}{a vector}

\item{type}{a vector}

\item{parent}{a vector}

\item{ext}{a vector}

\item{x}{object}
}
\value{
An S3 vector of class \verb{<ccnr>}
}
\description{
Construct a \code{ccnr} object
}
\examples{
x <- get_pin("ccn")

tibble::tibble(
  x = x,
  ccn = as_ccn(x),
  ccnr = as_ccnr(ccn))

y <- decode_ccnr(x) |>
  collapse::roworderv(c("entity", "state", "facility", "number")) |>
  collapse::rsplit(~entity)

tibble::tibble(
  parent0 = y$Unit$parent,
  parent1 = as_ccn(parent0),
  parent2 = as_ccnr(parent0)
)

y
}
