% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-ccnr.R
\name{ccnr}
\alias{ccnr}
\alias{new_ccnr}
\alias{is_ccnr}
\alias{as_ccnr}
\title{Construct a \code{ccnr} object}
\usage{
ccnr(
  ccn = character(),
  form = character(),
  state = character(),
  number = character(),
  type = character(),
  parent = character(),
  ext = character()
)

new_ccnr(
  ccn = character(),
  form = character(),
  state = character(),
  number = character(),
  type = character(),
  parent = character(),
  ext = character()
)

is_ccnr(x)

as_ccnr(x)
}
\arguments{
\item{ccn}{a vector}

\item{form}{a vector}

\item{state}{Passed on to methods}

\item{number}{a vector}

\item{type}{a vector}

\item{parent}{a vector}

\item{ext}{a vector}

\item{x}{object}
}
\value{
An S3 vector of class \verb{<ccnr>}
}
\description{
Construct a \code{ccnr} object
}
\examples{
x <- ccnr(
  c("670005", "12345E", NA, "05P001", "21T101", "02TA01", "10C0001062", "000001A001"),
  c("care", "erh", NA, "opo", "unit", "sub", "supp", "care"),
  c("67", "12", NA, "05", "21", "02", "10", "00"),
  c("0005", "345", NA, "001", "101", "01", "0001062", "0001"),
  c(NA, "E", NA, "P", "T", "T", "C", NA),
  c(NA, NA, NA, NA, NA, "A", NA, NA),
  c(NA, NA, NA, NA, NA, NA, NA, "A001"))
x
tibble::tibble(x, ccn = as_ccn(x))
tibble::as_tibble(vctrs::vec_data(x))
as_ccnr(get_pin("ccn"))
}
