% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{data_frame}
\alias{data_frame}
\alias{as_list}
\alias{as_data_frame}
\title{Convert to list/data.frame}
\usage{
as_list(x, ...)

as_data_frame(x, ...)
}
\arguments{
\item{x}{ccn object}

\item{...}{additional arguments}
}
\value{
list or data.frame
}
\description{
Convert to list/data.frame
}
\examples{
as_list(ccn("670055"))
as_list(ccn("05P001"))

c("670055", "05P001", "210101", "21T101", "21S101", "21U101",
  "01L008", "12345E", "10C0001062", "45D0634589", "21X0009807",
  "02TA01", "04SD38", "52TA05", "212026", "21SA26", "21TA26",
  "24T019A", "33S23401", "330027001") |>
purrr::map(ccn) |>
as_data_frame()
}
