% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-as_data_frame.R
\name{data_frame}
\alias{data_frame}
\title{Convert to a data.frame}
\usage{
data_frame(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
data.frame
}
\description{
Convert various codes to their associated names.
}
\examples{
list(
parse("670055"),
parse("21034E"),
parse("01L008"),
parse("01J008"),
parse("05P001"),
parse("21U101"),
parse("21TA26"),
parse("45D0634589")
) |>
  purrr::map(as.data.frame) |>
  purrr::list_rbind()
data_frame("670055")
data_frame("21034E")
data_frame("01L008")
data_frame("01J008")
data_frame("05P001")
data_frame("21U101")
data_frame("21TA26")
data_frame("45D0634589")
}
