% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\alias{decode.ccn}
\alias{decode.ccnr}
\title{Decode \code{ccn}/\code{ccnr} objects}
\usage{
decode(x, ...)

\method{decode}{ccn}(x, ...)

\method{decode}{ccnr}(x, ...)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}

\item{...}{Passed on to methods.}
}
\value{
An S3 vector of class \verb{<ccn>}
}
\description{
Decode \code{ccn}/\code{ccnr} objects
}
\examples{
x = get_pin("ccn")
decode(as_ccn(x))
decode(as_ccnr(x))
}
