% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decode CCNs}
\usage{
decode(x, ...)
}
\arguments{
\item{x}{character vector of codes to look up.}

\item{...}{additional arguments (not used).}
}
\value{
S7 object of class \code{Medicare}.
}
\description{
Convert various codes to their associated names.
}
\examples{
decode(slice("670055"))
decode(slice("123456"))
}
