% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decode a CCN}
\usage{
decode(x)
}
\arguments{
\item{x}{character vector of CCNs.}
}
\value{
list of CCN components.
}
\description{
Decode a CCN into its component parts.
}
\examples{
decode("67P055")
decode("670055")
decode("21-0101")
decode("21-T101")
decode("12345E")
decode("12C4567890")
decode("0-12C4567890")
}
