% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decode a CCN}
\usage{
decode(x)
}
\arguments{
\item{x}{A character vector of CCNs to decode. Must be either 6 or 10 characters long.}
}
\value{
A list with the decoded components of the CCN.
}
\description{
Decode a CCN into its component parts.
}
\examples{
decode("670055")
decode("21-0101")
decode("21-T101")
}
