% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decode CCNs}
\usage{
decode(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{Medicare}.
}
\description{
Convert various codes to their associated names.
}
\examples{
decode("670055")
decode("05P001")
decode("123456")
decode("21034E")
decode("12345F")
decode("210101")
decode("21T101")
decode("21S101")
decode("21U101")
decode("01L008")
decode("22B201")
decode("23E301")
decode("34F401")
decode("01J008")
decode("A5J508")
decode("21SA26")
decode("21TA26")
decode("02TA01")
decode("04SD38")
decode("52TA05")
decode("10C0001062")
decode("45D0634589")
decode("21X0009807")
decode("12C0001062")
decode("240019A")
decode("330125001")
decode("000000000")
decode("000000000000")
}
