% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-decode.R
\name{decode}
\alias{decode}
\title{Decode CMS Certification Numbers}
\usage{
decode(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{Medicare}.
}
\description{
Convert various codes to their associated names.
}
\examples{
x = list("670055", "21034E", "01L008", "01J008", "05P001", "21U101", "21TA26", "45D0634589")
purrr::map(x, decode)
}
