% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decode a CCN}
\usage{
decode(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{character vector of CCNs.}

\item{arg}{argument used for error messaging.}

\item{call}{environment used for error messaging.}
}
\value{
list of CCN components.
}
\description{
Decode a CCN into its component parts.
}
\examples{
decode("670055")
decode("21-0101")
decode("21-T101")
}
