% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R
\name{decode}
\alias{decode}
\title{Decode a CCN}
\usage{
decode(x)
}
\arguments{
\item{x}{character vector of CCNs.}
}
\value{
list of CCN components.
}
\description{
Decode a CCN into its component parts.
}
\examples{
decode("670055") # Medicare Provider
decode("05P001") # Medicare OPO
decode("12345E") # Emergency Hospital
decode("01L008") # Medicaid Only Provider

decode("210101")
decode("21T101")
decode("21S101")
decode("21U101")

decode("10C0001062") # Supplier ASC
decode("45D0634589") # Supplier CLIA
decode("21X0009807") # Supplier Portable X-Ray

decode("02TA01") # IPPS Excluded Unit
decode("04SD38")
decode("52TA05")

decode("212026") # Parent
decode("21SA26")
decode("21TA26")

decode("24T019A")
decode("33S23401")
decode("330027001")
}
