% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipps_excluded.R
\name{eipps}
\alias{eipps}
\alias{unit}
\alias{parent}
\title{IPPS-Excluded Hospital Units/Swing-Bed Approvals}
\usage{
unit(x)

eipps(x)

parent(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare("210101")
unit("21T101")
unit("21S101")
unit("21U101")

medicare("212026")
eipps("21SA26")
parent("21SA26")
eipps("21TA26")
parent("21TA26")

medicare("022001")
eipps("02TA01")
parent("02TA01")

medicare("043038")
eipps("04SD38")
parent("04SD38")

medicare("522005")
eipps("52TA05")
parent("52TA05")
}
