% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-emergency.R
\name{emergency}
\alias{emergency}
\title{Emergency Hospitals}
\usage{
emergency(x)
}
\arguments{
\item{x}{\verb{<character>} 6-position alphanumeric code.}
}
\value{
S7 object of class \verb{<Emergency>}.
}
\description{
CCNs assigned to Emergency Hospitals (both Federal and non-Federal)
that do not participate in the Medicare program.
}
\details{
Emergency hospitals are identified by a 6-position alphanumeric CCN.

Positions one and two identify the state in which the hospital is located.

Position six, an alpha character, identifies whether the hospital is Federal
(\code{F}) or non-Federal (\code{E}).

Positions three, four and five represent a sequence beginning with \code{001}.
This sequence identifies the order, by state, in which the CCN was issued.

For example, the CCN of the 34th emergency hospital in Maryland would be
\verb{21034E}. The Regional Office assigns the CCN in strict numerical sequence
without regard to the Federal or non-Federal status.
}
\examples{
emergency("21034E")
emergency("12345F")
emergency("11038F") # USAF Hospital (Moody)
}
