% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index_ccn}
\alias{index_ccn}
\title{Index a \code{ccn} object}
\usage{
index_ccn(x)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}
}
\value{
An S3 vector of class \verb{<ccn_index>}
}
\description{
Index a \code{ccn} object
}
\examples{
index_ccn(get_pin("ccn"))
}
