% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-STATES.R, R/lookups.R
\docType{data}
\name{state_codes}
\alias{state_codes}
\alias{state_names}
\alias{lookups}
\alias{emergency_hospital_name}
\alias{supplier_name}
\alias{state_name}
\alias{state_abbr}
\title{Lookups for Various Codes}
\format{
An object of class \code{character} of length 108.

An object of class \code{character} of length 59.
}
\usage{
state_codes

state_names

emergency_hospital_name(x)

supplier_name(x)

state_name(x)

state_abbr(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
LETTERS[emergency_hospital_name(LETTERS)]
LETTERS[supplier_name(LETTERS)]
state_name(state_codes)
state_abbr(state_names)
}
\keyword{internal}
