% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{lookups}
\alias{lookups}
\alias{get_emergency_abbr}
\alias{get_supplier_abbr}
\alias{get_state_abbr}
\alias{get_emergency_name}
\alias{get_supplier_name}
\alias{get_state_name}
\title{Lookups for Various Codes}
\usage{
get_emergency_abbr(x)

get_supplier_abbr(x)

get_state_abbr(x)

get_emergency_name(x)

get_supplier_name(x)

get_state_name(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
get_emergency_abbr(c("E", "F"))
get_supplier_abbr(c("C", "D", "X"))
get_state_abbr(ccn:::state_ccn_codes)

get_emergency_name(get_emergency_abbr(c("E", "F")))
get_supplier_name(get_supplier_abbr(c("C", "D", "X")))
get_state_name(get_state_abbr(ccn:::state_ccn_codes))
}
