% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{lookups}
\alias{lookups}
\alias{emergency_hospital_name}
\alias{supplier_name}
\alias{state_name}
\alias{state_abbr}
\title{Lookups for Various Codes}
\usage{
emergency_hospital_name(x)

supplier_name(x)

state_name(x)

state_abbr(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
emergency_hospital_name(LETTERS)
supplier_name(LETTERS)
state_name(c("00", "01", "83", "A0"))
state_abbr(c("Georgia", "Florida"))
}
