% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{lookups}
\alias{lookups}
\alias{get_emergency_abbr}
\alias{get_emergency_name}
\alias{get_supplier_abbr}
\alias{get_supplier_name}
\alias{get_caid_abbr}
\alias{get_caid_name}
\alias{get_ipps_abbr}
\alias{get_ipps_name}
\title{Lookups for Various Codes}
\usage{
get_emergency_abbr(x)

get_emergency_name(x)

get_supplier_abbr(x)

get_supplier_name(x)

get_caid_abbr(x)

get_caid_name(x)

get_ipps_abbr(x)

get_ipps_name(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
c("E", "F") |>
get_emergency_abbr() |>
print() |>
get_emergency_name()

c("C", "D", "X") |>
get_supplier_abbr() |>
print() |>
get_supplier_name()

c("A", "B", "E", "F", "G", "H", "J", "K", "L") |>
get_caid_abbr() |>
print() |>
get_caid_name()

c("M", "R", "S", "T", "U", "W", "Y", "Z") |>
get_ipps_abbr() |>
print() |>
get_ipps_name()
}
