% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaid.R
\name{medicaid}
\alias{medicaid}
\title{Medicaid-Only Providers}
\usage{
medicaid(x)
}
\arguments{
\item{x}{\verb{<character>} 6-position alphanumeric code.}
}
\value{
S7 object of class \verb{<Medicaid>}.
}
\description{
Title XIX-only providers are identified by a 6-digit alphanumeric CCN. The
first 2 digits identify the State in which the provider is located. The third
position, which is an alpha character, identifies the type of facility by
level or type of care being provided. The last 3 digits make up a sequential
number series beginning with \code{001}.
}
\examples{
medicaid("01L008")
medicaid("22B201")
medicaid("23E301")
medicaid("34F401")
medicaid("01J008")
medicaid("A5J508")
}
