% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaid.R
\name{medicaid}
\alias{medicaid}
\alias{medicaid_only_range}
\alias{medicaid_only_range_abbr}
\alias{medicaid_only_range_desc}
\alias{medicaid_only_sequence}
\alias{is_type_medicaid_only}
\alias{medicaid_only_type_abbr}
\alias{medicaid_only_type_desc}
\alias{medicaid_only_type}
\title{Medicaid-Only}
\usage{
medicaid_only_range(x)

medicaid_only_range_abbr(x)

medicaid_only_range_desc(x)

medicaid_only_sequence(x)

is_type_medicaid_only(x)

medicaid_only_type_abbr(x)

medicaid_only_type_desc(x)

medicaid_only_type(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicaid_only_range(150)
medicaid_only_range_abbr("100-199")
medicaid_only_range_desc("100-199")
medicaid_only_sequence("250")
is_type_medicaid_only(c("A", "J", "L"))
medicaid_only_type_abbr(c("A", "J", "L"))
medicaid_only_type_desc(c("A", "J", "L"))
medicaid_only_type("A")
}
