% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaid.R
\name{medicaid}
\alias{medicaid}
\title{Medicaid-Only Facilities}
\usage{
medicaid(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{Medicaid}.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicaid("01L008")
medicaid("22B201")
medicaid("23E301")
medicaid("34F401")
medicaid("01J008")
medicaid("A5J508")
}
