% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaid_facility.R
\name{medicaid_facility}
\alias{medicaid_facility}
\alias{new_medicaid_only_facility}
\title{Medicaid-Only Facilities}
\usage{
new_medicaid_only_facility(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
new_medicaid_only_facility("01L008")
new_medicaid_only_facility("22B201")
new_medicaid_only_facility("23E301")
new_medicaid_only_facility("34F401")
}
