% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicaid_hospital.R
\name{medicaid_only_hospital}
\alias{medicaid_only_hospital}
\title{Medicaid-Only Hospitals}
\usage{
medicaid_only_hospital(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicaid_only_hospital("01J008")
medicaid_only_hospital("A5J508")
}
