% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicaid_ranges}
\alias{medicaid_ranges}
\title{CCN Medicaid-Only Hospital Ranges}
\format{
\subsection{\code{medicaid_ranges}}{

A data frame with 7 rows and 5 columns:
\describe{
\item{start}{Medicaid-only hospital range start}
\item{end}{Medicaid-only hospital range end}
\item{range}{Medicaid-only hospital range}
\item{abbr}{Medicaid-only hospital abbreviation}
\item{desc}{Medicaid-only hospital description}
...
}
}
}
\source{
\url{https://www.cms.gov/regulations-and-guidance/guidance/manuals/downloads/som107c02.pdf}
}
\usage{
medicaid_ranges
}
\description{
\itemize{
\item Section: \strong{2779B}
\item Title: "CMS Certification Numbers for Medicaid Providers"
\item Pgs. 407-408
\item Rev. 123
\item Iss./Eff./Impl. 10-03-14
}
}
\keyword{internal}
