% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-medicare.R
\name{medicare}
\alias{medicare}
\title{Medicare Providers}
\usage{
medicare(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{Medicare}.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare("670055")
medicare("123456")
medicare("119802")
medicare("114036") # Greenleaf
medicare("110122") # SGMC
}
