% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare.R
\name{medicare}
\alias{medicare}
\alias{new_medicare}
\title{Medicare Providers}
\usage{
new_medicare(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
new_medicare("670055")
}
