% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode.R, R/medicare.R
\name{print.ccn::Medicare}
\alias{print.ccn::Medicare}
\alias{medicare}
\title{Medicare Provider}
\usage{
\method{print}{`ccn::Medicare`}(x, ...)

medicare(x)
}
\arguments{
\item{x}{character vector of codes to look up.}

\item{...}{additional arguments (not used).}
}
\value{
S7 object of class \code{Medicare}.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare("670055")
medicare("123456")
}
