% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicare_ranges}
\alias{medicare_ranges}
\title{CCN Medicare Facility Ranges}
\format{
\subsection{\code{medicare_ranges}}{

A data frame with 39 rows and 3 columns:
\describe{
\item{range}{Facility range}
\item{abbr}{Facility abbreviation}
\item{desc}{Facility description}
...
}
}
}
\source{
\url{https://www.cms.gov/regulations-and-guidance/guidance/manuals/downloads/som107c02.pdf}
}
\usage{
medicare_ranges
}
\description{
From Section 2779A1: CCN for Medicare Providers (pgs. 405-406)
(Rev. 226; Issued: 12-06-24; Effective: 12-06-24; Implementation: 12-06-24)
}
\keyword{datasets}
