% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medicare_ranges}
\alias{medicare_ranges}
\title{CCN Medicare Facility Ranges}
\format{
\subsection{\code{medicare_ranges}}{

A data frame with 39 rows and 3 columns:
\describe{
\item{range}{Medicare facility range}
\item{abbr}{Medicare facility abbreviation}
\item{desc}{Medicare facility description}
...
}
}
}
\source{
\url{https://www.cms.gov/regulations-and-guidance/guidance/manuals/downloads/som107c02.pdf}
}
\usage{
medicare_ranges
}
\description{
\itemize{
\item Section \strong{2779A1}
\item Title: "CCN for Medicare Providers"
\item Pgs. 405-406
\item Rev. 226
\item Iss./Eff./Impl. 12-06-24
}
}
\keyword{internal}
