% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergency.R, R/opo.R
\name{new_emergency}
\alias{new_emergency}
\alias{opo}
\alias{opo_sequence}
\alias{opo_type}
\alias{new_opo}
\title{Organ Procurement Organizations (OPO)}
\usage{
new_emergency(x)

opo_sequence(x)

opo_type(x)

new_opo(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Functions for identifying and describing Organ Procurement Organizations (OPO) based on their codes.
}
\examples{
new_opo("05P001")
}
