% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opo.R
\name{opo}
\alias{opo}
\alias{is_type_opo}
\alias{opo_abbr}
\alias{opo_desc}
\alias{opo_sequence}
\alias{opo_type}
\title{Organ Procurement Organizations (OPO)}
\usage{
is_type_opo(x)

opo_abbr(x)

opo_desc(x)

opo_sequence(x)

opo_type(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Functions for identifying and describing Organ Procurement Organizations (OPO) based on their codes.
}
\examples{
is_type_opo("P")
opo_abbr(c("P", "X"))
opo_desc("P")
opo_sequence("001")
opo_type("P")
}
