% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opo.R
\name{organ}
\alias{organ}
\title{Organ Procurement Organizations (OPO)}
\usage{
organ(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
S7 object of class \code{Organ}.
}
\description{
Organ Procurement Organizations (OPO) based on their codes.
}
\examples{
organ("05P001")
}
