% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{parent}
\alias{parent}
\alias{get_parent_prefix}
\alias{get_parent_abbr}
\alias{get_unit_sequence}
\alias{get_parent_ccn}
\title{IPPS Excluded Unit Lookups}
\usage{
get_parent_prefix(x)

get_parent_abbr(x)

get_unit_sequence(x)

get_parent_ccn(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
x <- c("02TA01", "04SD38", "52TA05")
get_parent_prefix(x)
get_parent_abbr(x)
get_unit_sequence(x)
get_parent_ccn(x)
}
