% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{parent}
\alias{parent}
\title{Unit Parents}
\usage{
parent(x, ...)
}
\arguments{
\item{x}{character vector of codes to look up.}

\item{...}{additional arguments (not used).}
}
\value{
S7 object of class \code{Medicare}.
}
\description{
Convert various codes to their associated names.
}
\examples{
parent("21AE26")
parent("21SE26")
parent("21SB26")
parent("21SD26")
}
