% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\title{Parse CMS Certification Numbers}
\usage{
parse(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
parse("670055")
parse("05P001")
parse("12345E")
parse("10C0001062")
parse("A5J508")
parse("21T101")
parse("02TA01")
parse("240019A")
parse("330125001")
parse("000000000")
parse("00000000000000")
}
