% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\alias{parse_medicare_provider}
\alias{parse_medicare_opo}
\alias{parse_emergency_hospital}
\alias{parse_medicare_supplier}
\alias{parse_medicaid_only_hospital}
\alias{parse_medicaid_only_facility}
\alias{parse_eipps_provider}
\alias{parse_eipps_unit}
\title{Parse CMS Certification Numbers}
\usage{
parse_medicare_provider(x)

parse_medicare_opo(x)

parse_emergency_hospital(x)

parse_medicare_supplier(x)

parse_medicaid_only_hospital(x)

parse_medicaid_only_facility(x)

parse_eipps_provider(x)

parse_eipps_unit(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
parse_medicare_provider("670055")
parse_medicare_opo("05P001")
parse_emergency_hospital("12345E")
parse_medicare_supplier("10C0001062")
parse_medicaid_only_hospital("A5J508")
parse_medicaid_only_facility("23E301")
parse_eipps_provider("21T101")
parse_eipps_unit("02TA01")
}
