% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\alias{parse_excluded}
\alias{parse_medicaid_only}
\alias{parse_opo}
\alias{parse_medicare}
\alias{parse_emergency}
\alias{parse_supplier}
\alias{parse_excluded_unit}
\title{Parse CMS Certification Numbers}
\usage{
parse_excluded(x)

parse_medicaid_only(x)

parse_opo(x)

parse_medicare(x)

parse_emergency(x)

parse_supplier(x)

parse_excluded_unit(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
parse_medicare("670055")
parse_opo("05P001")
parse_emergency("12345E")
parse_supplier("10C0001062")
parse_excluded("21T101")
parse_excluded_unit("02TA01")
}
