% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse}
\alias{parse}
\alias{parse_medicare}
\alias{parse_opo}
\alias{parse_emergency}
\alias{parse_supplier}
\alias{parse_medicaid}
\alias{parse_unit}
\alias{parse_subunit}
\title{Parse CMS Certification Numbers}
\usage{
parse_medicare(x)

parse_opo(x)

parse_emergency(x)

parse_supplier(x)

parse_medicaid(x)

parse_unit(x)

parse_subunit(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
parse_medicare("670055")
parse_opo("05P001")
parse_emergency("12345E")
parse_supplier("10C0001062")
parse_medicaid("A5J508")
parse_unit("21T101")
parse_subunit("02TA01")
}
