% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates}
\alias{predicates}
\alias{type_medicaid}
\alias{type_excluded}
\alias{type_emergency}
\alias{type_supplier}
\alias{type_opo}
\title{Logical Predicates}
\usage{
type_medicaid(x)

type_excluded(x)

type_emergency(x)

type_supplier(x)

type_opo(x)
}
\arguments{
\item{x}{character vector of codes to test.}
}
\value{
logical vector indicating if codes are of the specified type.
}
\description{
Logical tests for various code types
}
\examples{
x <- LETTERS
x[type_medicaid(x)]
x[type_excluded(x)]
x[type_emergency(x)]
x[type_supplier(x)]
x[type_opo(x)]
}
