% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates}
\alias{predicates}
\alias{is_medicaid_facility_type}
\alias{is_ipps_excluded_type}
\alias{is_emergency_hospital_code}
\alias{is_supplier_code}
\title{Logical Predicates}
\usage{
is_medicaid_facility_type(x)

is_ipps_excluded_type(x)

is_emergency_hospital_code(x)

is_supplier_code(x)
}
\arguments{
\item{x}{character vector of codes to test.}
}
\value{
logical vector indicating if codes are of the specified type.
}
\description{
Logical tests for various code types
}
\examples{
is_medicaid_facility_type(LETTERS)
is_ipps_excluded_type(LETTERS)
is_emergency_hospital_code(LETTERS)
is_supplier_code(LETTERS)
}
