% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{predicates}
\alias{predicates}
\alias{is_type_medicaid}
\alias{is_type_excluded}
\alias{is_type_unit}
\alias{is_type_emergency}
\alias{is_type_supplier}
\alias{is_type_opo}
\title{Logical Predicates}
\usage{
is_type_medicaid(x)

is_type_excluded(x)

is_type_unit(x)

is_type_emergency(x)

is_type_supplier(x)

is_type_opo(x)
}
\arguments{
\item{x}{character vector of codes to test.}
}
\value{
logical vector indicating if codes are of the specified type.
}
\description{
Logical tests for various code types
}
\examples{
x <- LETTERS
x[is_type_medicaid(x)]
x[is_type_excluded(x)]
x[is_type_unit(x)]
x[is_type_emergency(x)]
x[is_type_supplier(x)]
x[is_type_opo(x)]
}
