% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges}
\alias{ranges}
\alias{get_care_range}
\alias{get_care_range_abbr}
\alias{get_care_range_name}
\alias{get_caid_range}
\alias{get_caid_range_abbr}
\alias{get_caid_range_name}
\alias{get_opo_range}
\alias{get_emergency_range}
\alias{get_supplier_range}
\alias{get_parent_prefix}
\alias{get_parent_abbr}
\alias{get_unit_sequence}
\title{Facility Ranges}
\usage{
get_care_range(x)

get_care_range_abbr(x)

get_care_range_name(x)

get_caid_range(x)

get_caid_range_abbr(x)

get_caid_range_name(x)

get_opo_range(x)

get_emergency_range(x)

get_supplier_range(x)

get_parent_prefix(x)

get_parent_abbr(x)

get_unit_sequence(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
get_care_range(c("0055", "5232", "9999")) |>
 print() |>
 get_care_range_abbr() |>
 print() |>
 get_care_range_name()

get_caid_range(c("055", "232", "599")) |>
 print() |>
 get_caid_range_abbr() |>
 print() |>
 get_caid_range_name()

get_supplier_range("9999999")
get_opo_range("50")
}
