% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges}
\alias{ranges}
\alias{get_care_range}
\alias{get_care_range_abbr}
\alias{get_care_range_name}
\alias{get_caid_range}
\alias{get_caid_range_abbr}
\alias{get_caid_range_name}
\alias{get_opo_range}
\alias{get_opo_range_abbr}
\alias{get_opo_range_name}
\title{Lookups for Facility Ranges}
\usage{
get_care_range(x)

get_care_range_abbr(x)

get_care_range_name(x)

get_caid_range(x)

get_caid_range_abbr(x)

get_caid_range_name(x)

get_opo_range(x)

get_opo_range_abbr(x)

get_opo_range_name(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
get_care_range(c("0055", "5232", "9999")) |>
 print() |>
 get_care_range_abbr() |>
 print() |>
 get_care_range_name()

get_caid_range(c("055", "232", "599")) |>
 print() |>
 get_caid_range_abbr() |>
 print() |>
 get_caid_range_name()

 get_opo_range(c("01", "12", "99")) |>
 print() |>
 get_opo_range_abbr() |>
 print() |>
 get_opo_range_name()
}
