% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges}
\alias{ranges}
\alias{range_medicare_opo}
\alias{range_emergency}
\alias{range_medicare_supplier}
\alias{range_medicaid_only_abbr}
\alias{range_medicaid_only_desc}
\alias{range_medicaid_only}
\alias{range_medicare_abbr}
\alias{range_medicare_desc}
\alias{range_medicare}
\title{Facility Ranges}
\usage{
range_medicare_opo(x)

range_emergency(x)

range_medicare_supplier(x)

range_medicaid_only_abbr(x)

range_medicaid_only_desc(x)

range_medicaid_only(x)

range_medicare_abbr(x)

range_medicare_desc(x)

range_medicare(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
range_medicare(c("0055", "5232", "9999")) |>
 print() |>
 range_medicare_abbr() |>
 print() |>
 range_medicare_desc()

range_medicaid_only(c("055", "232", "599")) |>
 print() |>
 range_medicaid_only_abbr() |>
 print() |>
 range_medicaid_only_desc()

range_medicare_supplier("9999999")
range_emergency("50")
range_medicare_opo("50")
}
