% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode-ranges}
\alias{recode-ranges}
\alias{recode_medicare_range}
\alias{recode_medicaid_range}
\title{Recode CMS Facility Ranges}
\usage{
recode_medicare_range(x, as = c("abbr", "full", "range"))

recode_medicaid_range(x, as = c("abbr", "full", "range"))
}
\arguments{
\item{x}{\verb{<int>} vector of CCN facility ranges.}

\item{as}{\verb{<chr>} format to return; one of:
\itemize{
\item \code{"abbr"}: facility range abbreviation (default)
\item \code{"full"}: facility range full description
\item \code{"range"}: facility range text representation
}}
}
\value{
\verb{<chr>} vector of CMS facility range information
}
\description{
Recode CMS Facility Ranges
}
\examples{
tibble::tibble(
  x = c(1L, 2099L),
  abbr = recode_medicare_range(x, "abbr"),
  range = recode_medicare_range(x, "range"),
  facility = recode_medicare_range(x, "full")
)

tibble::tibble(
  x = c(1L, 2099L),
  abbr = recode_medicaid_range(x, "abbr"),
  range = recode_medicaid_range(x, "range"),
  facility = recode_medicaid_range(x, "full")
)
}
