% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_region}
\alias{recode_region}
\title{Recode CMS Region Codes}
\usage{
recode_region(x, as = c("number", "roman", "abbr", "full"))
}
\arguments{
\item{x}{\verb{<chr>} vector of CCN state codes.}

\item{as}{\verb{<chr>} format to return; one of:
\itemize{
\item "number": CMS region number (default)
\item "roman": CMS region number as a roman numeral
\item "abbr": CMS region's office location abbreviation
\item "full": CMS region's office location name
}}
}
\value{
\verb{<chr>} vector of CMS region names/abbreviations
}
\description{
Recode CMS Region Codes
}
\examples{
tibble::tibble(
  state = c("00", "01", "A5", NA),
  number = recode_region(state, "number"),
  roman = recode_region(state, "roman"),
  abbr = recode_region(state, "abbr"),
  full = recode_region(state, "full")
)
}
