% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions}
\alias{regions}
\title{CMS Regional Office Codes}
\format{
\subsection{\code{regions}}{

A data frame with 59 rows and 4 columns:
\describe{
\item{ro}{Region number}
\item{region}{Region roman numeral}
\item{state}{State abbreviation}
\item{city}{Regional office city}
...
}
}
}
\usage{
regions
}
\description{
CMS Regional Office Codes
}
\keyword{internal}
