% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice}
\alias{slice}
\alias{slice_medicare}
\alias{slice_organ}
\alias{slice_emergency}
\alias{slice_supplier}
\alias{slice_medicaid}
\alias{slice_unit}
\alias{slice_subunit}
\title{Slice CMS Certification Numbers}
\usage{
slice_medicare(x)

slice_organ(x)

slice_emergency(x)

slice_supplier(x)

slice_medicaid(x)

slice_unit(x)

slice_subunit(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
slice_medicare("670055")
slice_organ("05P001")
slice_emergency("12345E")
slice_supplier("10C0001062")
slice_medicaid("A5J508")
slice_unit("21T101")
slice_subunit("02TA01")
}
