% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_codes.R
\name{state_code}
\alias{state_code}
\title{State Code Lookup
Convert state codes to standard state names.}
\usage{
state_code(x)
}
\arguments{
\item{x}{character vector of state codes.}
}
\value{
character vector of state names.
}
\description{
State Code Lookup
Convert state codes to standard state names.
}
\examples{
state_code(c("00", "01", "83", "A0"))
}
