% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_codes}
\alias{state_codes}
\title{CCN State Codes}
\format{
\subsection{\code{state_codes}}{

A data frame with 108 rows and 3 columns:
\describe{
\item{code}{State code}
\item{abbr}{State abbreviation}
\item{name}{State full name}
...
}
}
}
\source{
\url{https://www.cms.gov/regulations-and-guidance/guidance/manuals/downloads/som107c02.pdf}
}
\usage{
state_codes
}
\description{
From Section 2779A1: CCN for Medicare Providers (pgs. 403-404)
(Rev. 226; Issued: 12-06-24; Effective: 12-06-24; Implementation: 12-06-24)
}
\keyword{datasets}
