% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{CCN State Codes}
\format{
\subsection{\code{states}}{

A data frame with 139 rows and 3 columns:
\describe{
\item{code}{State code}
\item{abbr}{State abbreviation}
\item{full}{State full name}
...
}
}
}
\source{
Updated (12-06-24) Provider State Codes - CMS Manual System Pub. 100-07: \href{https://www.cms.gov/files/document/r226soma.pdf}{State Operations Manual Revisions, section 2779A1}
}
\usage{
states
}
\description{
CCN State Codes
}
\keyword{internal}
