% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R, R/data.R
\docType{data}
\name{states}
\alias{states}
\alias{state}
\title{State Codes}
\format{
\subsection{\code{states}}{

A data frame with 108 rows and 3 columns:
\describe{
\item{code}{State code}
\item{abbr}{State abbreviation}
\item{name}{State full name}
...
}
}
}
\usage{
state(x)

states
}
\arguments{
\item{x}{character vector of state codes to look up.}
}
\value{
S7 object of class \code{State}.
}
\description{
Convert state codes to their abbreviations or full names.
}
\examples{
state("A5XXX")
state("00")
}
\keyword{internal}
