% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{CCN State Codes}
\format{
\subsection{\code{states}}{

A data frame with 108 rows and 3 columns:
\describe{
\item{code}{State code}
\item{abbr}{State abbreviation}
\item{name}{State full name}
...
}
}
}
\usage{
states
}
\description{
CCN State Codes
}
\keyword{internal}
