% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{states}
\alias{states}
\alias{get_state_abbr}
\alias{get_state_name}
\alias{is_state_code}
\title{State Lookups}
\usage{
get_state_abbr(x)

get_state_name(x)

is_state_code(x)
}
\arguments{
\item{x}{character vector of state codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert state codes to their abbreviations or full names.
}
\examples{
all(is_state_code(ccn:::state_codes))

get_state_abbr(ccn:::state_codes) |>
print() |>
get_state_name()
}
