% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/State.R
\name{states}
\alias{states}
\alias{is_state_code}
\alias{get_state_code}
\alias{get_state_abbr}
\alias{get_state_name}
\title{State Codes, Abbreviations, and Names}
\usage{
is_state_code(x)

get_state_code(x)

get_state_abbr(x)

get_state_name(x)
}
\arguments{
\item{x}{character vector of state codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert state codes to their abbreviations or full names.
}
\examples{
get_state_code(c("A5XXX", "14XXXXX"))

(x <- sample(ccn:::state_codes, 5L))

all(is_state_code(x))

(x <- get_state_abbr(x))

get_state_name(x)
}
