% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunit.R
\name{subunit}
\alias{subunit}
\title{IPPS-Excluded Hospital Sub-Units}
\usage{
subunit(x)
}
\arguments{
\item{x}{\if{html}{\out{<character>}} scalar.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
subunit("21SA26")
subunit("21TA26")
subunit("02TA01")
subunit("04SD38")
subunit("52TA05")
}
