% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunit.R
\name{subunit}
\alias{subunit}
\alias{parent}
\title{IPPS-Excluded Hospital Sub-Units}
\usage{
subunit(x)

parent(x)
}
\arguments{
\item{x}{\if{html}{\out{<character>}} scalar.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare("212026")
subunit("21SA26")
parent("21SA26")
subunit("21TA26")
parent("21TA26")

medicare("022001")
subunit("02TA01")
parent("02TA01")

medicare("043038")
subunit("04SD38")
parent("04SD38")

medicare("522005")
subunit("52TA05")
parent("52TA05")
}
