% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{supplier}
\alias{supplier}
\title{Suppliers}
\usage{
supplier(x)
}
\arguments{
\item{x}{\verb{<character>} 10-position alphanumeric code.}
}
\value{
S7 object of class \verb{<Supplier>}.
}
\description{
Suppliers that are paid by Part B carriers have a 10-digit alphanumeric CCN.
The first 2 digits identify the State in which the supplier is located. The
third digit is an alpha character that identifies the type of facility. The
remaining 7 digits are the unique facility identifier.

The RO assigns the following alpha-characters in the third position as
indicated:
\itemize{
\item \code{C} - Ambulatory Surgical Centers
\item \code{D} - Clinical Laboratory Improvement Amendments of 1988 (CLIA) Laboratories
\item \code{X} - Portable X-Ray Facilities
}
}
\examples{
supplier("10C0001062")
supplier("45D0634589")
supplier("21X0009807")
supplier("12C0001062")
}
