% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{supplier}
\alias{supplier}
\alias{new_supplier}
\title{Medicare Supplier}
\usage{
new_supplier(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
new_supplier("10C0001062")
new_supplier("45D0634589")
new_supplier("21X0009807")
new_supplier("12C0001062")
}
