% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{supplier}
\alias{supplier}
\alias{is_supplier_type}
\alias{is_supplier}
\alias{supplier_sequence}
\alias{supplier_type_abbr}
\alias{supplier_type_desc}
\alias{supplier_type}
\title{Medicare Supplier}
\usage{
is_supplier_type(x)

is_supplier(x)

supplier_sequence(x)

supplier_type_abbr(x)

supplier_type_desc(x)

supplier_type(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
is_supplier("10C0001062")
supplier_sequence("10C0001062")
is_supplier_type(c("C", "D", "X"))
supplier_type_abbr(c("C", "D", "X"))
supplier_type_desc(c("C", "D", "X"))
}
