% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.R
\name{unit}
\alias{unit}
\title{IPPS-Excluded Hospital Units}
\usage{
unit(x)
}
\arguments{
\item{x}{character vector of codes to look up.}
}
\value{
character vector of names associated with codes.
}
\description{
Convert various codes to their associated names.
}
\examples{
medicare("210101")
unit("21T101")
unit("21S101")
unit("21U101")
}
