% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.R
\name{unit}
\alias{unit}
\title{Hospital Units}
\usage{
unit(x)
}
\arguments{
\item{x}{\verb{<character>} 6-position alphanumeric code.}
}
\value{
S7 object of class \verb{<Unit>}.
}
\description{
An alpha character in the third position of a CCN identifies either hospitals
with swing-bed approval, rehabilitation units, or psychiatric units excluded
from IPPS payment.

The first 2 digits identify the State in which the provider is located. The
third position (which is alpha) identifies the type of unit or swing-bed
designation.

\strong{The last 3 digits must match the last 3 digits of the parent provider.}
}
\examples{
medicare("210101")
unit("21T101")
unit("21S101")
unit("21U101")
unit("28Z348")
medicare("281348")
}
