% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R, R/index.R
\name{vctrs}
\alias{vctrs}
\alias{ccn}
\alias{ccnr}
\alias{get_index}
\alias{get_index.ccn}
\alias{get_index.ccnr}
\title{Construct a \code{ccn}/\code{ccnr} object}
\usage{
ccn(x = character())

ccnr(
  ccn = character(),
  entity = character(),
  state = character(),
  number = character(),
  type = character()
)

get_index(x)

\method{get_index}{ccn}(x, ...)

\method{get_index}{ccnr}(x, ...)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}

\item{ccn}{\verb{<chr>} vector of ccns}

\item{entity}{\verb{<chr>} vector of ccn entity types}

\item{state}{\verb{<chr>} vector of state codes}

\item{number}{\verb{<chr>} vector of sequence numbers}

\item{type}{\verb{<chr>} vector of facility types}

\item{...}{description}
}
\value{
An S3 vector of class \verb{<ccn>} or \verb{<ccnr>}
}
\description{
Construct a \code{ccn}/\code{ccnr} object
}
\examples{
x = get_pin("ccn")
y = ccn(x)
get_index(y)
z = as_ccnr(x)
get_index(z)

vctrs::vec_c(x[1:5], y[100:150], x[1:50], z[100:150])
tibble::tibble(x = x, ccn = ccn(x))
}
