% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{vctrs}
\alias{vctrs}
\alias{ccn}
\alias{get_index}
\alias{get_index.ccn}
\alias{ccnr}
\title{Construct a \code{ccn}/\code{ccnr} object}
\usage{
ccn(x = character())

get_index(x)

\method{get_index}{ccn}(x, ...)

ccnr(
  entity = character(),
  ccn = character(),
  state = character(),
  number = character(),
  type = character()
)
}
\arguments{
\item{x}{\verb{<chr>} A vector of CCNs.}

\item{...}{description}

\item{entity}{\verb{<chr>} vector of ccn entity types}

\item{ccn}{\verb{<chr>} vector of ccns}

\item{state}{\verb{<chr>} vector of state codes}

\item{number}{\verb{<chr>} vector of sequence numbers}

\item{type}{\verb{<chr>} vector of facility types}
}
\value{
An S3 vector of class \verb{<ccn>} or \verb{<ccnr>}
}
\description{
Construct a \code{ccn}/\code{ccnr} object
}
\examples{
x = get_pin("ccn")
y = ccn(x)
get_index(y)

vctrs::vec_c(x[1:5], y[100:150], x[1:50])
tibble::tibble(x = x, ccn = ccn(x))

ccnr(
  entity = "Medicare",
  ccn = "001234",
  state = "00",
  number = "1234",
  type = NA_character_
 )
}
