% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn_rcrd.R
\name{vec_proxy_equal.ccnr}
\alias{vec_proxy_equal.ccnr}
\title{Comparison}
\usage{
\method{vec_proxy_equal}{ccnr}(x, ...)
}
\arguments{
\item{x}{A vector x.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Dispatch methods to support \code{\link[vctrs:vec_proxy_equal]{vctrs::vec_proxy_equal()}}.
}
