% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccn.R
\name{vec_ptype2.ccn}
\alias{vec_ptype2.ccn}
\title{Coercion}
\usage{
\method{vec_ptype2}{ccn}(x, y, ..., x_arg = "", y_arg = "")
}
\arguments{
\item{x, y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}

\item{x_arg, y_arg}{Argument names for \code{x} and \code{y}. These are used
in error messages to inform the user about the locations of
incompatible types (see \code{\link[vctrs:stop_incompatible_type]{stop_incompatible_type()}}).}
}
\description{
Double dispatch methods to support \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}.
}
\keyword{internal}
